@@ -1,11 +1,11 @@
-from collections.abc import Mapping, Container
+import importlib
 import logging
 import os
 import pkgutil
+import sys
 from collections import OrderedDict
 from inspect import isfunction
 from operator import itemgetter
-import sys
 from types import SimpleNamespace
 from typing import Union
 
@@ -14,22 +14,16 @@ import h5py
 import numpy as np
 import torch
 import yaml
-from omegaconf import OmegaConf, DictConfig
+from omegaconf import DictConfig, OmegaConf
 
 log = logging.getLogger(__name__)
 
 
 def load_yaml(filename: Union[str, os.PathLike]) -> dict:
     """Simple wrapper around yaml.load to load yaml files as dictionaries"""
-    with open(filename, 'r') as f:
+    with open(filename, "r") as f:
         dictionary = yaml.load(f, Loader=yaml.Loader)
     return dictionary
-
-
-# def load_config(filename: Union[str, os.PathLike]) -> DictConfig:
-#     """ loads a yaml file as dictionary and converts to an omegaconf DictConfig """
-#     dictionary = load_yaml(filename)
-#     return OmegaConf.create(dictionary)
 
 
 def get_minimum_learning_rate(optimizer):
@@ -38,19 +32,21 @@ def get_minimum_learning_rate(optimizer):
     """
     min_lr = 1e9
     for i, param_group in enumerate(optimizer.param_groups):
-        lr = param_group['lr']
+        lr = param_group["lr"]
         if lr < min_lr:
             min_lr = lr
-    return (min_lr)
+    return min_lr
 
 
-def load_checkpoint(model,
-                    optimizer,
-                    checkpoint_file: Union[str, os.PathLike],
-                    config: dict,
-                    overwrite_args: bool = False,
-                    distributed: bool = False):
-    """"Reload model and optimizer weights from a checkpoint.pt file
+def load_checkpoint(
+    model,
+    optimizer,
+    checkpoint_file: Union[str, os.PathLike],
+    config: dict,
+    overwrite_args: bool = False,
+    distributed: bool = False,
+):
+    """ "Reload model and optimizer weights from a checkpoint.pt file
     Args:
         model: instance of torch.nn.Module class
         optimizer: instance of torch.optim.Optimizer class (ADAM, SGDM, etc.)
@@ -62,29 +58,29 @@ def load_checkpoint(model,
         optimizer: optimizer with recent history of gradients
         config: depending on overwrite_args, input or reloaded hyperparameter dictionary
     """
-    log.info('Reloading model from {}...'.format(checkpoint_file))
+    log.info("Reloading model from {}...".format(checkpoint_file))
     model, optimizer_dict, _, new_args = load_state(model, checkpoint_file, distributed=distributed)
-    if type(new_args) != dict:
-        new_config = vars(new_args)
+    if isinstance(new_args, dict):
+        new_config = new_args
     else:
         new_config = new_args
     try:
         optimizer.load_state_dict(optimizer_dict)
     except Exception as e:
-        log.exception('Trouble loading optimizer state dict--might have requires-grad' \
-                      'for different parameters: {}'.format(e))
-        log.warning('Not loading optimizer state.')
+        log.exception(
+            "Trouble loading optimizer state dict--might have requires-gradfor different parameters: {}".format(e)
+        )
+        log.warning("Not loading optimizer state.")
     if overwrite_args:
         config = new_config
 
     return model, optimizer, config
 
 
-def load_weights(model,
-                 checkpoint_file: Union[str, os.PathLike],
-                 distributed: bool = False,
-                 device: torch.device = None):
-    """"Reload model weights from a checkpoint.pt file
+def load_weights(
+    model, checkpoint_file: Union[str, os.PathLike], distributed: bool = False, device: torch.device = None
+):
+    """ "Reload model weights from a checkpoint.pt file
     Args:
         model: instance of torch.nn.Module class
         checkpoint_file: path to checkpoint.pt
@@ -100,7 +96,7 @@ def load_weights(model,
 
 
 def checkpoint(model, rundir: Union[str, os.PathLike], epoch: int, args=None):
-    """"
+    """ "
     Args:
         model: instance of torch.nn.Module class
         rundir: directory to  save checkpoint.pt to
@@ -110,28 +106,28 @@ def checkpoint(model, rundir: Union[str, os.PathLike], epoch: int, args=None):
     if args is not None:
         if isinstance(args, DictConfig):
             args = OmegaConf.to_container(args)
-        if type(args) != dict:
+        if not isinstance(args, dict):
             args = vars(args)
-    fname = 'checkpoint.pt'
+    fname = "checkpoint.pt"
     fullfile = os.path.join(rundir, fname)
     # note: I used to save the optimizer dict as well, but this was confusing in terms of keeping track of learning
     # rates, making sure the same keys were in the optimizer dict even when you've done something like change
     # the size of the final layer of the NN (for different number of classes). I've kept the optimizer field for
     # backwards compatibility, but this should not be used
-    state = {'epoch': epoch, 'state_dict': model.state_dict(), 'optimizer': None, 'hyperparameters': args}
+    state = {"epoch": epoch, "state_dict": model.state_dict(), "optimizer": None, "hyperparameters": args}
     torch.save(state, fullfile)
 
 
 def save_two_stream(model, rundir: Union[os.PathLike, str], config: dict = None, epoch: int = None) -> None:
-    """ Saves a two-stream model to disk. Saves spatial and flow feature extractors in their own directories """
+    """Saves a two-stream model to disk. Saves spatial and flow feature extractors in their own directories"""
     assert os.path.isdir(rundir)
     assert isinstance(model, torch.nn.Module)
-    spatialdir = os.path.join(rundir, 'spatial')
+    spatialdir = os.path.join(rundir, "spatial")
     if not os.path.isdir(spatialdir):
         os.makedirs(spatialdir)
     checkpoint(model.spatial_classifier, spatialdir, epoch, config)
 
-    flow_classifier_dir = os.path.join(rundir, 'flow')
+    flow_classifier_dir = os.path.join(rundir, "flow")
     if not os.path.isdir(flow_classifier_dir):
         os.makedirs(flow_classifier_dir)
     checkpoint(model.flow_classifier, flow_classifier_dir, epoch, config)
@@ -139,14 +135,16 @@ def save_two_stream(model, rundir: Union[os.PathLike, str], config: dict = None,
     checkpoint(model, rundir, epoch, config)
 
 
-def save_hidden_two_stream(model, rundir: Union[os.PathLike, str], config: dict = None, epoch: int = None) -> None:
-    """ Saves a hidden two-stream model to disk. Saves flow generator in a separate directory """
+def save_hidden_two_stream(
+    model, rundir: Union[os.PathLike, str], config: Union[dict, None] = None, epoch: int = None
+) -> None:
+    """Saves a hidden two-stream model to disk. Saves flow generator in a separate directory"""
     assert os.path.isdir(rundir)
     assert isinstance(model, torch.nn.Module)
-    flowdir = os.path.join(rundir, 'flow_generator')
+    flowdir = os.path.join(rundir, "flow_generator")
     if not os.path.isdir(flowdir):
         os.makedirs(flowdir)
-    if type(config) == DictConfig:
+    if isinstance(config, DictConfig):
         config = OmegaConf.to_container(config)
     checkpoint(model.flow_generator, flowdir, epoch, config)
     save_two_stream(model, rundir, config, epoch)
@@ -159,22 +157,22 @@ def save_dict_to_yaml(dictionary: dict, filename: Union[str, bytes, os.PathLike]
         filename: file to save dict to. Should end in .yaml
     """
     if os.path.isfile(filename):
-        log.debug('File {} already exists, overwriting...'.format(filename))
+        log.debug("File {} already exists, overwriting...".format(filename))
     if isinstance(dictionary, DictConfig):
         dictionary = OmegaConf.to_container(dictionary)
-    with open(filename, 'w') as f:
+    with open(filename, "w") as f:
         yaml.dump(dictionary, f, default_flow_style=False)
 
 
 def tensor_to_np(tensor: Union[torch.Tensor, np.ndarray]) -> np.ndarray:
     """Simple function for turning pytorch tensor into numpy ndarray"""
-    if type(tensor) == np.ndarray:
+    if isinstance(tensor, np.ndarray):
         return tensor
     return tensor.cpu().detach().numpy()
 
 
 def in_this_dir(abs_path: Union[str, os.PathLike]) -> dict:
-    """ Gets list of files in a subdirectory and returns information about it.
+    """Gets list of files in a subdirectory and returns information about it.
     Designed to be a drop-in replacement for MATLAB's `dir` command :P
     Args:
         abs_path: absolute path to a directory
@@ -182,20 +180,20 @@ def in_this_dir(abs_path: Union[str, os.PathLike]) -> dict:
         contents: dictionary with keys 'name', 'isdir', and 'bytes', containing the name, whether or not the file is
             a directory, and the filesize in bytes of all files in the directory
     """
-    backslashes = strfind(abs_path, '\\')
+    backslashes = strfind(abs_path, "\\")
     if len(backslashes) > 1 and backslashes[0] != -1:
-        abs_path.replace('\\', '/')
+        abs_path.replace("\\", "/")
     # contents contains list of filenames or directory names
     filenames = os.listdir(abs_path)
     contents = []
     for name in filenames:
         content = {}
-        content['name'] = name
-        content['isdir'] = os.path.isdir(os.path.join(abs_path, name))
-        content['bytes'] = os.path.getsize(os.path.join(abs_path, name))
+        content["name"] = name
+        content["isdir"] = os.path.isdir(os.path.join(abs_path, name))
+        content["bytes"] = os.path.getsize(os.path.join(abs_path, name))
         contents.append(content)
     # sort by name!
-    contents = sorted(contents, key=itemgetter('name'))
+    contents = sorted(contents, key=itemgetter("name"))
     return contents
 
 
@@ -210,7 +208,7 @@ def get_datadir_from_paths(paths, dataset):
             datadir = v
             found = True
     if not found:
-        raise ValueError('couldn' 't find dataset: {}'.format(dataset))
+        raise ValueError("couldnt find dataset: {}".format(dataset))
     return datadir
 
 
@@ -238,51 +236,44 @@ def load_state_from_dict(model, state_dict):
     model_dict = model.state_dict()
     pretrained_dict = {}
     for k, v in state_dict.items():
-        if 'criterion' in k:
+        if "criterion" in k:
             # we might have parameters from the loss function in our loaded weights. we don't want to reload these;
             # we will specify them for whatever we are currently training.
             continue
         if k not in model_dict:
-            log.warning('{} not found in model dictionary'.format(k))
+            log.warning("{} not found in model dictionary".format(k))
         else:
             if model_dict[k].size() != v.size():
-                log.warning('{} has different size: pretrained:{} model:{}'.format(k, v.size(), model_dict[k].size()))
+                log.warning("{} has different size: pretrained:{} model:{}".format(k, v.size(), model_dict[k].size()))
             else:
-                log.debug('Successfully loaded: {}'.format(k))
+                log.debug("Successfully loaded: {}".format(k))
                 pretrained_dict[k] = v
 
     model_dict.update(pretrained_dict)
-    # only_in_model_dict = {k:v for k,v in state_dict.items() if k in model_dict}
-    # model_dict.update(only_in_model_dict)
     # load the state dict, only for layers of same name, shape, size, etc.
     model.load_state_dict(model_dict, strict=True)
-    return (model)
+    return model
 
 
 def load_state_dict_from_file(weights_file, distributed: bool = False):
-    state = torch.load(weights_file, map_location='cpu')
-    # except RuntimeError as e:
-    #     log.exception(e)
-    #     log.info('loading onto cpu...')
-    #     state = torch.load(weights_file, map_location='cpu')
+    state = torch.load(weights_file, map_location="cpu")
 
-    is_pure_weights = not 'epoch' in list(state.keys())
+    is_pure_weights = "epoch" not in list(state.keys())
     # load params
     if is_pure_weights:
         state_dict = state
         start_epoch = 0
     else:
-        start_epoch = state['epoch']
-        state_dict = state['state_dict']
-        optimizer_dict = None  # state['optimizer']
+        start_epoch = state["epoch"]
+        state_dict = state["state_dict"]
 
     first_key = next(iter(state_dict.items()))[0]
-    trained_on_dataparallel = first_key[:7] == 'module.'
+    trained_on_dataparallel = first_key[:7] == "module."
 
     if distributed and not trained_on_dataparallel:
         new_state_dict = OrderedDict()
         for k, v in state_dict.items():
-            name = 'module.' + k
+            name = "module." + k
             new_state_dict[name] = v
         state_dict = new_state_dict
     # if it was trained on multi-gpu, remove the 'module.' before variable names
@@ -294,19 +285,19 @@ def load_state_dict_from_file(weights_file, distributed: bool = False):
         state_dict = new_state_dict
     # sometimes I have the encoder in a structure called 'model', which means
     # all weights have 'model.' prepended
-    model_prepended = first_key[:6] == 'model.'
+    model_prepended = first_key[:6] == "model."
     if model_prepended:
         new_state_dict = OrderedDict()
         for k, v in state_dict.items():
-            if k[:6] == 'model.':
+            if k[:6] == "model.":
                 name = k[6:]
             else:
                 name = k
             new_state_dict[name] = v
         state_dict = new_state_dict
     if not is_pure_weights:
-        if 'hyperparameters' in list(state.keys()):
-            args = state['hyperparameters']
+        if "hyperparameters" in list(state.keys()):
+            args = state["hyperparameters"]
         else:
             args = None
     else:
@@ -316,7 +307,7 @@ def load_state_dict_from_file(weights_file, distributed: bool = False):
 
 
 def load_state(model, weights_file: Union[str, os.PathLike], device: torch.device = None, distributed: bool = False):
-    """"Reload model and optimizer weights from a checkpoint.pt file.
+    """ "Reload model and optimizer weights from a checkpoint.pt file.
 
     TODO: refactor this loading for pytorch 1.4+. This was written many versions ago
 
@@ -332,13 +323,12 @@ def load_state(model, weights_file: Union[str, os.PathLike], device: torch.devic
         args: SimpleNamespace containing hyperparameters
             TODO: change args to a config dictionary
     """
-    # fullfile = os.path.join(model_dir,run_dir, fname)
     # state is a dictionary
     # Keys:
     #  epoch: final epoch number from training
     #  state_dict: weights
     #  args: hyperparameters
-    log.info('loading from checkpoint file {}...'.format(weights_file))
+    log.info("loading from checkpoint file {}...".format(weights_file))
 
     state_dict, start_epoch, args = load_state_dict_from_file(weights_file, distributed=distributed)
     # LOAD PARAMS
@@ -354,9 +344,16 @@ def print_gpus():
     """
     n_gpus = torch.cuda.device_count()
     for i in range(n_gpus):
-        print('GPU %d %s: Compute Capability %d.%d, Mem:%f' %
-              (i, torch.cuda.get_device_name(i), int(torch.cuda.get_device_capability(i)[0]),
-               int(torch.cuda.get_device_capability(i)[1]), torch.cuda.max_memory_allocated(i)))
+        print(
+            "GPU %d %s: Compute Capability %d.%d, Mem:%f"
+            % (
+                i,
+                torch.cuda.get_device_name(i),
+                int(torch.cuda.get_device_capability(i)[0]),
+                int(torch.cuda.get_device_capability(i)[1]),
+                torch.cuda.max_memory_allocated(i),
+            )
+        )
 
 
 class Normalizer:
@@ -376,10 +373,13 @@ class Normalizer:
         mean: mean of input data. For images, should have 2 or 3 channels
         std: standard deviation of input data
     """
-    def __init__(self,
-                 mean: Union[list, np.ndarray, torch.Tensor] = None,
-                 std: Union[list, np.ndarray, torch.Tensor] = None,
-                 clamp: bool = True):
+
+    def __init__(
+        self,
+        mean: Union[list, np.ndarray, torch.Tensor] = None,
+        std: Union[list, np.ndarray, torch.Tensor] = None,
+        clamp: bool = True,
+    ):
         """Constructor for Normalizer class.
         Args:
             mean: mean of input data. Should have 3 channels (for R,G,B) or 2 (for X,Y) in the optical flow case
@@ -389,28 +389,28 @@ class Normalizer:
         # make sure that if you have a mean, you also have a std
         # XOR
         has_mean, has_std = mean is None, std is None
-        assert (not has_mean ^ has_std)
+        assert not has_mean ^ has_std
 
         self.mean = self.process_inputs(mean)
         self.std = self.process_inputs(std)
         # prevent divide by zero, but only change values if it's close to 0 already
         if self.std is not None:
-            assert (self.std.min() > 0)
+            assert self.std.min() > 0
             self.std[self.std < 1e-8] += 1e-8
-        log.debug('Normalizer created with mean {} and std {}'.format(self.mean, self.std))
+        log.debug("Normalizer created with mean {} and std {}".format(self.mean, self.std))
         self.clamp = clamp
 
-    def process_inputs(self, inputs: Union[torch.Tensor, np.ndarray]):
+    def process_inputs(self, inputs: Union[torch.Tensor, np.ndarray, list, None]):
         """Deals with input mean and std.
         Converts to tensor if necessary. Reshapes to [length, 1, 1] for pytorch broadcasting.
         """
         if inputs is None:
-            return (inputs)
-        if type(inputs) == list:
+            return inputs
+        if isinstance(inputs, list):
             inputs = np.array(inputs).astype(np.float32)
-        if type(inputs) == np.ndarray:
+        if isinstance(inputs, np.ndarray):
             inputs = torch.from_numpy(inputs)
-        assert (type(inputs) == torch.Tensor)
+        assert isinstance(inputs, torch.Tensor)
         inputs = inputs.float()
         C = inputs.shape[0]
         inputs = inputs.reshape(C, 1, 1)
@@ -436,7 +436,7 @@ class Normalizer:
         elif tensor.ndim == 5:
             N, C, T, H, W = tensor.shape
         else:
-            raise ValueError('Tensor input to normalizer of unknown shape: {}'.format(tensor.shape))
+            raise ValueError("Tensor input to normalizer of unknown shape: {}".format(tensor.shape))
 
         t_d = tensor.device
         if t_d != self.mean.device:
@@ -452,7 +452,7 @@ class Normalizer:
             # this code simply repeats the mean T times, so it's
             # [R_mean, G_mean, B_mean, R_mean, G_mean, ... etc]
             n_repeats = C / c
-            assert (int(n_repeats) == n_repeats)
+            assert int(n_repeats) == n_repeats
             n_repeats = int(n_repeats)
             repeats = tuple([n_repeats] + [1 for i in range(self.mean.ndim - 1)])
             self.mean = self.mean.repeat((repeats))
@@ -557,37 +557,25 @@ def flow_img_to_flow(img: np.ndarray, max_flow: Union[int, float] = 10) -> np.nd
     return np.dstack((dX, dY))
 
 
-# def encode_flow_img(flow, maxflow=10):
-#     im = flow_to_img_lrcn(flow, max_flow=maxflow)
-#     # print(im.shape)
-#     ret, bytestring = cv2.imencode('.jpg', im)
-#     return (bytestring)
-
-# def decode_flow_img(bytestring, maxflow=10):
-#     im = cv2.imdecode(bytestring, 1)
-#     flow = flow_img_to_flow(im, max_flow=maxflow)
-#     return (flow)
-
-
 def module_to_dict(module, exclude=[], get_function=False):
-    """ Converts functions in a module to a dictionary. Useful for loading model types into a dictionary """
+    """Converts functions in a module to a dictionary. Useful for loading model types into a dictionary"""
     module_dict = {}
     for x in dir(module):
         submodule = getattr(module, x)
         # print(x, submodule)
         func = isfunction(submodule) if get_function else not isfunction(submodule)
-        if (func and x not in exclude and submodule not in exclude):
+        if func and x not in exclude and submodule not in exclude:
             module_dict[x] = submodule
     return module_dict
 
 
 def get_models_from_module(module, get_function=False):
-    """ Hacky function for getting a dictionary of model: initializer from a module """
+    """Hacky function for getting a dictionary of model: initializer from a module"""
     models = {}
     for importer, modname, ispkg in pkgutil.iter_modules(module.__path__):
-        # print("Found submodule %s (is a package: %s)" % (modname, ispkg))
-        total_name = module.__name__ + '.' + modname
-        this_module = __import__(total_name)
+        total_name = module.__name__ + "." + modname
+        # Import the module and get its attributes
+        importlib.import_module(total_name)
         submodule = getattr(module, modname)
         # module
         this_dict = module_to_dict(submodule, get_function=get_function)
@@ -598,7 +586,7 @@ def get_models_from_module(module, get_function=False):
 
 
 def load_feature_extractor_components(model, checkpoint_file: Union[str, os.PathLike], component: str, device=None):
-    """ Loads individual component from a hidden two-stream model checkpoint
+    """Loads individual component from a hidden two-stream model checkpoint
 
     Parameters
     ----------
@@ -616,42 +604,26 @@ def load_feature_extractor_components(model, checkpoint_file: Union[str, os.Path
     model: nn.Module
         pytorch model with loaded weights
     """
-    if component == 'spatial':
-        key = 'spatial_classifier' + '.'
-    elif component == 'flow':
-        key = 'flow_classifier' + '.'
-    elif component == 'fusion':
-        key = 'fusion.'
+    if component == "spatial":
+        key = "spatial_classifier" + "."
+    elif component == "flow":
+        key = "flow_classifier" + "."
+    elif component == "fusion":
+        key = "fusion."
     else:
-        raise ValueError('component not one of spatial or flow: {}'.format(component))
-    # directory = os.path.dirname(checkpoint_file)
-    # subdir = os.path.join(directory, component)
-    # log.info('device: {}'.format(device))
-    log.info('loading component {} from file {}'.format(component, checkpoint_file))
+        raise ValueError("component not one of spatial or flow: {}".format(component))
+
+    log.info("loading component {} from file {}".format(component, checkpoint_file))
 
     state_dict, _, _ = load_state_dict_from_file(checkpoint_file)
 
-    # state = torch.load(checkpoint_file, map_location=device)
-    # state_dict = state['state_dict']
-    params = {k.replace(key, ''): v for k, v in state_dict.items() if k.startswith(key)}
-    # import pdb; pdb.set_trace()
-    model = load_state_from_dict(model, params)
-    # import pdb; pdb.set_trace()
-    # if not os.path.isdir(subdir):
-    #     log.warning('{} directory not found in {}'.format(component, directory))
-    #     state = torch.load(checkpoint_file, map_location=device)
-    #     state_dict = state['state_dict']
-    #     params = {k.replace(key, ''): v for k, v in state_dict.items() if k.startswith(key)}
-    #     # import pdb; pdb.set_trace()
-    #     model = load_state_from_dict(model, params)
-    # else:
-    #     sub_checkpoint = os.path.join(subdir, 'checkpoint.pt')
-    #     model, _, _, _ = load_state(model, sub_checkpoint, device=device)
-    return model
+    params = {k.replace(key, ""): v for k, v in state_dict.items() if k.startswith(key)}
+
+    return load_state_from_dict(model, params)
 
 
 def get_subfiles(root: Union[str, bytes, os.PathLike], return_type: str = None) -> list:
-    """ Helper function to get a list of files of certain type from a directory
+    """Helper function to get a list of files of certain type from a directory
 
     Parameters
     ----------
@@ -667,21 +639,21 @@ def get_subfiles(root: Union[str, bytes, os.PathLike], return_type: str = None) 
     files: list
         list of absolute paths of sub-files
     """
-    assert (return_type in [None, 'any', 'file', 'directory'])
+    assert return_type in [None, "any", "file", "directory"]
     files = os.listdir(root)
     files.sort()
     files = [os.path.join(root, i) for i in files]
-    if return_type is None or return_type == 'any':
+    if return_type is None or return_type == "any":
         pass
-    elif return_type == 'file':
+    elif return_type == "file":
         files = [i for i in files if os.path.isfile(i)]
-    elif return_type == 'directory':
+    elif return_type == "directory":
         files = [i for i in files if os.path.isdir(i)]
     return files
 
 
 def print_hdf5(h5py_obj, level=-1, print_full_name: bool = False, print_attrs: bool = True) -> None:
-    """ Prints the name and shape of datasets in a H5py HDF5 file.
+    """Prints the name and shape of datasets in a H5py HDF5 file.
     Parameters
     ----------
     h5py_obj: [h5py.File, h5py.Group]
@@ -697,22 +669,23 @@ def print_hdf5(h5py_obj, level=-1, print_full_name: bool = False, print_attrs: b
     -------
     None
     """
+
     def is_group(f):
-        return type(f) == h5py._hl.group.Group
+        return isinstance(f, h5py._hl.group.Group)
 
     def is_dataset(f):
-        return type(f) == h5py._hl.dataset.Dataset
+        return isinstance(f, h5py._hl.dataset.Dataset)
 
     def print_level(level, n_spaces=5) -> str:
         if level == -1:
-            return ''
-        prepend = '|' + ' ' * (n_spaces - 1)
+            return ""
+        prepend = "|" + " " * (n_spaces - 1)
         prepend *= level
-        tree = '|' + '-' * (n_spaces - 2) + ' '
+        tree = "|" + "-" * (n_spaces - 2) + " "
         return prepend + tree
 
-    if isinstance(h5py_obj, str) or isinstance(h5py_obj, os.PathLike):
-        with h5py.File(h5py_obj, 'r') as f:
+    if isinstance(h5py_obj, (str, os.PathLike)):
+        with h5py.File(h5py_obj, "r") as f:
             print_hdf5(f)
             return
 
@@ -720,63 +693,29 @@ def print_hdf5(h5py_obj, level=-1, print_full_name: bool = False, print_attrs: b
         entry = h5py_obj[key]
         name = entry.name if print_full_name else os.path.basename(entry.name)
         if is_group(entry):
-            print('{}{}'.format(print_level(level), name))
+            print("{}{}".format(print_level(level), name))
             print_hdf5(entry, level + 1, print_full_name=print_full_name)
         elif is_dataset(entry):
             shape = entry.shape
             dtype = entry.dtype
-            print('{}{}: {} {}'.format(print_level(level), name, shape, dtype))
+            print("{}{}: {} {}".format(print_level(level), name, shape, dtype))
     if level == -1:
         if print_attrs:
-            print('attrs: ')
-
-
-#
-# def deep_getsizeof(o, ids):
-#     """Find the memory footprint of a Python object
-#
-#     This is a recursive function that drills down a Python object graph
-#     like a dictionary holding nested dictionaries with lists of lists
-#     and tuples and sets.
-#
-#     The sys.getsizeof function does a shallow size of only. It counts each
-#     object inside a container as pointer only regardless of how big it
-#     really is.
-#
-#     :param o: the object
-#     :param ids:
-#     :return:
-#     """
-#     d = deep_getsizeof
-#     if id(o) in ids:
-#         return 0
-#
-#     r = sys.getsizeof(o)
-#     ids.add(id(o))
-#
-#     if isinstance(o, str):
-#         return r
-#
-#     if isinstance(o, Mapping):
-#         return r + sum(d(k, ids) + d(v, ids) for k, v in o.iteritems())
-#
-#     if isinstance(o, Container):
-#         return r + sum(d(x, ids) for x in o)
-#
-#     return r
+            print("attrs: ")
 
 
 def print_top_largest_variables(local_call, num: int = 20):
-    def sizeof_fmt(num, suffix='B'):
-        ''' by Fred Cirera,  https://stackoverflow.com/a/1094933/1870254, modified'''
-        for unit in ['', 'Ki', 'Mi', 'Gi', 'Ti', 'Pi', 'Ei', 'Zi']:
+    def sizeof_fmt(num, suffix="B"):
+        """by Fred Cirera,  https://stackoverflow.com/a/1094933/1870254, modified"""
+        for unit in ["", "Ki", "Mi", "Gi", "Ti", "Pi", "Ei", "Zi"]:
             if abs(num) < 1024.0:
                 return "%3.1f %s%s" % (num, unit, suffix)
             num /= 1024.0
-        return "%.1f %s%s" % (num, 'Yi', suffix)
+        return "%.1f %s%s" % (num, "Yi", suffix)
 
-    for name, size in sorted(((name, sys.getsizeof(value)) for name, value in local_call.items()),
-                             key=lambda x: -x[1])[:10]:
+    for name, size in sorted(((name, sys.getsizeof(value)) for name, value in local_call.items()), key=lambda x: -x[1])[
+        :10
+    ]:
         print("{:>30}: {:>8}".format(name, sizeof_fmt(size)))
 
 
@@ -788,7 +727,7 @@ def get_hparams_from_cfg(cfg, hparams):
 def get_dotted_from_cfg(cfg, dotted):
     # cfg: DictConfig
     # dotted: string parameter name. can be nested. e.g. 'tune.hparams.feature_extractor.dropout_p.min'
-    key_list = dotted.split('.')
+    key_list = dotted.split(".")
 
     cfg_chunk = cfg.get(key_list[0])
     for i in range(1, len(key_list)):
@@ -798,34 +737,33 @@ def get_dotted_from_cfg(cfg, dotted):
 
 
 def get_best_epoch_from_weightfile(weightfile: Union[str, os.PathLike]) -> int:
-    """parses a checkpoint like epoch=15.ckpt to find the number 15
-    """
+    """parses a checkpoint like epoch=15.ckpt to find the number 15"""
     basename = os.path.basename(weightfile)
     # in the previous version of deepethogram, load the last checkpoint
-    if basename.endswith('.pt'):
+    if basename.endswith(".pt"):
         return -1
-    assert basename.endswith('.ckpt')
+    assert basename.endswith(".ckpt")
     basename = os.path.splitext(basename)[0]
 
     # if weightfile is the "last"
-    if 'last' in basename:
+    if "last" in basename:
         return -1
 
-    components = basename.split('-')
+    components = basename.split("-")
     component = components[0]
-    assert component.startswith('epoch')
-    best_epoch = component.split('=')[1]
+    assert component.startswith("epoch")
+    best_epoch = component.split("=")[1]
     return int(best_epoch)
 
 
 def remove_nans_and_infs(array: np.ndarray, set_value: float = 0.0) -> np.ndarray:
-    """ Simple function to remove nans and infs from a numpy array """
+    """Simple function to remove nans and infs from a numpy array"""
     bad_indices = np.logical_or(np.isinf(array), np.isnan(array))
     array[bad_indices] = set_value
     return array
 
 
-def get_run_files_from_weights(weightfile: Union[str, os.PathLike], metrics_prefix='classification') -> dict:
+def get_run_files_from_weights(weightfile: Union[str, os.PathLike], metrics_prefix="classification") -> dict:
     """from model weights, gets the configuration for that model and its metrics file
 
     Parameters
@@ -839,15 +777,15 @@ def get_run_files_from_weights(weightfile: Union[str, os.PathLike], metrics_pref
         config_file: path to config file
         metrics_file: path to metrics file
     """
-    loaded_config_file = os.path.join(os.path.dirname(weightfile), 'config.yaml')
+    loaded_config_file = os.path.join(os.path.dirname(weightfile), "config.yaml")
     if not os.path.isfile(loaded_config_file):
         # weight file should be at most one-subdirectory-down from rundir
-        loaded_config_file = os.path.join(os.path.dirname(os.path.dirname(weightfile)), 'config.yaml')
-        assert os.path.isfile(loaded_config_file), 'no associated config file for weights! {}'.format(weightfile)
+        loaded_config_file = os.path.join(os.path.dirname(os.path.dirname(weightfile)), "config.yaml")
+        assert os.path.isfile(loaded_config_file), "no associated config file for weights! {}".format(weightfile)
 
-    metrics_file = os.path.join(os.path.dirname(weightfile), f'{metrics_prefix}_metrics.h5')
+    metrics_file = os.path.join(os.path.dirname(weightfile), f"{metrics_prefix}_metrics.h5")
     if not os.path.isfile(metrics_file):
-        metrics_file = os.path.join(os.path.dirname(os.path.dirname(weightfile)), f'{metrics_prefix}_metrics.h5')
-        assert os.path.isfile(metrics_file), 'no associated metrics file for weights! {}'.format(weightfile)
+        metrics_file = os.path.join(os.path.dirname(os.path.dirname(weightfile)), f"{metrics_prefix}_metrics.h5")
+        assert os.path.isfile(metrics_file), "no associated metrics file for weights! {}".format(weightfile)
 
     return dict(config_file=loaded_config_file, metrics_file=metrics_file)
