@@ -7,9 +7,9 @@
 # from deepethogram.sequence.inference import sequence_inference
 import importlib.util
 
-spec = importlib.util.find_spec('hydra')
+spec = importlib.util.find_spec("hydra")
 if spec is not None:
-    raise ValueError('Hydra installation found. Please run pip uninstall hydra-core: {}'.format(spec))
+    raise ValueError("Hydra installation found. Please run pip uninstall hydra-core: {}".format(spec))
 # try:
 #     import hydra
 # except Exception as e:
