@@ -1,7 +1,7 @@
 import os
 from typing import Union
 
-from omegaconf import OmegaConf, DictConfig
+from omegaconf import DictConfig, OmegaConf
 
 import deepethogram
 from deepethogram import projects
@@ -15,25 +15,25 @@ def config_string_to_path(config_path: Union[str, os.PathLike], string: str) -> 
     config_path : Union[str, os.PathLike]
         absolute path to deepethogram/deepethogram/conf directory
     string : str
-        name of configuration. 
+        name of configuration.
 
     Returns
     -------
     str
         Absolute path to configuration default file
-        
+
     Examples
     --------
     >>> config_string_to_path('path/to/deepethogram/deepethogram/conf', 'tune/feature_extractor')
     'path/to/deepethogram/deepethogram/conf/tune/feature_extractor.yaml'
-        
+
     """
-    fullpath = os.path.join(config_path, *string.split('/')) + '.yaml'
-    assert os.path.isfile(fullpath), f'{fullpath} not found'
+    fullpath = os.path.join(config_path, *string.split("/")) + ".yaml"
+    assert os.path.isfile(fullpath), f"{fullpath} not found"
     return fullpath
 
 
-def load_config_by_name(string: str, config_path: Union[str, os.PathLike] = None) -> DictConfig:
+def load_config_by_name(string: str, config_path: Union[str, os.PathLike, None] = None) -> DictConfig:
     """Loads a default configuration by name
 
     Parameters
@@ -47,7 +47,7 @@ def load_config_by_name(string: str, config_path: Union[str, os.PathLike] = None
     -------
     DictConfig
         Configuration loaded from YAML file
-        
+
     Examples
     --------
     >>> load_config_by_name('model/feature_extractor')
@@ -73,50 +73,52 @@ def load_config_by_name(string: str, config_path: Union[str, os.PathLike] = None
     """
 
     if config_path is None:
-        config_path = os.path.join(os.path.dirname(deepethogram.__file__), 'conf')
+        config_path = os.path.join(os.path.dirname(deepethogram.__file__), "conf")
 
     path = config_string_to_path(config_path, string)
     return OmegaConf.load(path)
 
 
-def make_config(project_path: Union[str, os.PathLike],
-                config_list: list,
-                run_type: str,
-                model: str,
-                use_command_line: bool = False,
-                preset: str = None,
-                debug: bool = False) -> DictConfig:
-    """Makes a configuration for model training or inference. 
-    
-    A list of default configurations are composed into one single cfg. From the project path, the project configuration
-    is found and loaded. If a preset is specified either in the config_list or in the project config, load "preset" 
-    parameters. 
+def make_config(
+    project_path: Union[str, os.PathLike],
+    config_list: list,
+    run_type: str,
+    model: str,
+    use_command_line: bool = False,
+    preset: str = None,
+    debug: bool = False,
+) -> DictConfig:
+    """Makes a configuration for model training or inference.
 
-    Order of composition: 
+    A list of default configurations are composed into one single cfg. From the project path, the project configuration
+    is found and loaded. If a preset is specified either in the config_list or in the project config, load "preset"
+    parameters.
+
+    Order of composition:
     1. Defaults
     2. Preset
     3. Project configuration
     4. Command line
-    
-    This means if you specify the value of a parameter (say, dropout probability) in multiple places, the last one 
+
+    This means if you specify the value of a parameter (say, dropout probability) in multiple places, the last one
     (highest number in above list) will be chosen. This means we can specify a default dropout (0.25); for your project,
-    you can specify a new default in your project_config (e.g. 0.5). For an experiment, you can use the commmand line 
-    to set `feature_extractor.dropout_p=0.75`. If its in all 3 places, the command line "wins" and the actual dropout is 
-    0.75. 
+    you can specify a new default in your project_config (e.g. 0.5). For an experiment, you can use the commmand line
+    to set `feature_extractor.dropout_p=0.75`. If its in all 3 places, the command line "wins" and the actual dropout is
+    0.75.
 
     Parameters
     ----------
     project_path : Union[str, os.PathLike]
         Path to deepethogram project. Should contain: project_config.yaml, models directory, DATA directory
     config_list : list
-        List of string names of default configurations. Each of them is the name of a file or sub-file in the 
-        deepethogram/conf directory. 
+        List of string names of default configurations. Each of them is the name of a file or sub-file in the
+        deepethogram/conf directory.
     run_type : str
         Train, inference, or gui
     model : str
         feature_extractor, flow_generator, or sequence
     use_command_line : bool, optional
-        If True, command line arguments are parsed and composed into the 
+        If True, command line arguments are parsed and composed into the
     preset : str, optional
         One of deg_f, deg_m, deg_s, by default None
     debug : bool, optional
@@ -127,7 +129,6 @@ def make_config(project_path: Union[str, os.PathLike],
     DictConfig
         [description]
     """
-    # config_path = os.path.join(os.path.dirname(deepethogram.__file__), 'conf')
 
     user_cfg = projects.get_config_from_path(project_path)
 
@@ -136,22 +137,21 @@ def make_config(project_path: Union[str, os.PathLike],
     # then, append the user config
     # then, the command line args
     # so if we specify a preset and manually change, say, the feature extractor architecture, we can do that
-    if 'preset' in user_cfg:
-        config_list.append('preset/' + user_cfg.preset)
+    if "preset" in user_cfg:
+        config_list.append("preset/" + user_cfg.preset)
 
     if use_command_line:
         command_line_cfg = OmegaConf.from_cli()
-        if 'preset' in command_line_cfg:
-            config_list.append('preset/' + command_line_cfg.preset)
+        if "preset" in command_line_cfg:
+            config_list.append("preset/" + command_line_cfg.preset)
 
     # add this option so we can add a preset programmatically
     if preset is not None:
-        assert preset in ['deg_f', 'deg_m', 'deg_s']
-        config_list.append('preset/' + preset)
+        assert preset in ["deg_f", "deg_m", "deg_s"]
+        config_list.append("preset/" + preset)
 
     if debug:
-        config_list.append('debug')
-    # config_files = [config_string_to_path(config_path, i) for i in config_list]
+        config_list.append("debug")
 
     cfgs = [load_config_by_name(i) for i in config_list]
 
@@ -161,7 +161,7 @@ def make_config(project_path: Union[str, os.PathLike],
     else:
         cfg = OmegaConf.merge(*cfgs, user_cfg)
 
-    cfg.run = {'type': run_type, 'model': model}
+    cfg.run = {"type": run_type, "model": model}
     return cfg
 
 
@@ -178,9 +178,9 @@ def make_flow_generator_train_cfg(project_path: Union[str, os.PathLike], **kwarg
     DictConfig
         flow generator config
     """
-    config_list = ['config', 'augs', 'train', 'model/flow_generator']
-    run_type = 'train'
-    model = 'flow_generator'
+    config_list = ["config", "augs", "train", "model/flow_generator"]
+    run_type = "train"
+    model = "flow_generator"
 
     cfg = make_config(project_path=project_path, config_list=config_list, run_type=run_type, model=model, **kwargs)
 
@@ -200,9 +200,9 @@ def make_feature_extractor_train_cfg(project_path: Union[str, os.PathLike], **kw
     DictConfig
         feature extractor train config
     """
-    config_list = ['config', 'augs', 'train', 'model/flow_generator', 'model/feature_extractor']
-    run_type = 'train'
-    model = 'feature_extractor'
+    config_list = ["config", "augs", "train", "model/flow_generator", "model/feature_extractor"]
+    run_type = "train"
+    model = "feature_extractor"
 
     cfg = make_config(project_path=project_path, config_list=config_list, run_type=run_type, model=model, **kwargs)
 
@@ -222,9 +222,9 @@ def make_feature_extractor_inference_cfg(project_path: Union[str, os.PathLike], 
     DictConfig
         feature extractor inference config
     """
-    config_list = ['config', 'augs', 'model/feature_extractor', 'model/flow_generator', 'inference', 'postprocessor']
-    run_type = 'inference'
-    model = 'feature_extractor'
+    config_list = ["config", "augs", "model/feature_extractor", "model/flow_generator", "inference", "postprocessor"]
+    run_type = "inference"
+    model = "feature_extractor"
 
     cfg = make_config(project_path=project_path, config_list=config_list, run_type=run_type, model=model, **kwargs)
 
@@ -244,9 +244,9 @@ def make_sequence_train_cfg(project_path: Union[str, os.PathLike], **kwargs) -> 
     DictConfig
         sequence train config
     """
-    config_list = ['config', 'model/feature_extractor', 'train', 'model/sequence']
-    run_type = 'train'
-    model = 'sequence'
+    config_list = ["config", "model/feature_extractor", "train", "model/sequence"]
+    run_type = "train"
+    model = "sequence"
 
     cfg = make_config(project_path=project_path, config_list=config_list, run_type=run_type, model=model, **kwargs)
 
@@ -266,9 +266,9 @@ def make_sequence_inference_cfg(project_path: Union[str, os.PathLike], **kwargs)
     DictConfig
         sequence inference config
     """
-    config_list = ['config', 'augs', 'model/feature_extractor', 'model/sequence', 'inference']
-    run_type = 'inference'
-    model = 'sequence'
+    config_list = ["config", "augs", "model/feature_extractor", "model/sequence", "inference"]
+    run_type = "inference"
+    model = "sequence"
 
     cfg = make_config(project_path=project_path, config_list=config_list, run_type=run_type, model=model, **kwargs)
 
@@ -288,9 +288,9 @@ def make_postprocessing_cfg(project_path: Union[str, os.PathLike], **kwargs) -> 
     DictConfig
         postprocessing config
     """
-    config_list = ['config', 'model/sequence', 'postprocessor']
-    run_type = 'inference'
-    model = 'sequence'
+    config_list = ["config", "model/sequence", "postprocessor"]
+    run_type = "inference"
+    model = "sequence"
 
     cfg = make_config(project_path=project_path, config_list=config_list, run_type=run_type, model=model, **kwargs)
 
