@@ -1,30 +1,29 @@
-import bisect
-from collections import deque
 import logging
 import os
-import pprint
 import random
-import warnings
-from functools import partial
-from typing import Union, Tuple
+from collections import deque
+from typing import Tuple, Union
 
 import h5py
 import numpy as np
-from omegaconf import DictConfig
-import pandas as pd
 import torch
-from opencv_transforms import transforms
+from omegaconf import DictConfig
 from torch.utils import data
 from vidio import VideoReader
 
 # from deepethogram.dataloaders import log
 from deepethogram import projects
 from deepethogram.data.augs import get_cpu_transforms
-from deepethogram.data.utils import purge_unlabeled_elements_from_records, get_video_metadata, extract_metadata, \
-    find_labelfile, read_all_labels, get_split_from_records, remove_invalid_records_from_split_dictionary, \
-    make_loss_weight, fix_label
-from deepethogram.data.keypoint_utils import load_dlcfile, interpolate_bad_values, stack_features_in_time, \
-    expand_features_sturman
+from deepethogram.data.keypoint_utils import expand_features_sturman, interpolate_bad_values, load_dlcfile
+from deepethogram.data.utils import (
+    fix_label,
+    get_split_from_records,
+    get_video_metadata,
+    make_loss_weight,
+    purge_unlabeled_elements_from_records,
+    read_all_labels,
+    remove_invalid_records_from_split_dictionary,
+)
 from deepethogram.file_io import read_labels
 
 log = logging.getLogger(__name__)
@@ -32,22 +31,24 @@ log = logging.getLogger(__name__)
 
 # https://pytorch.org/docs/stable/data.html
 class VideoIterable(data.IterableDataset):
-    """Highly optimized Dataset for running inference on videos. 
-    
-    Features: 
+    """Highly optimized Dataset for running inference on videos.
+
+    Features:
         - Data is only read sequentially
         - Each frame is only read once
         - The input video is divided into NUM_WORKERS segments. Each worker reads its segment in parallel
-        - Each clip is read with stride = 1. If sequence_length==3, the first clips would be frames [0, 1, 2], 
+        - Each clip is read with stride = 1. If sequence_length==3, the first clips would be frames [0, 1, 2],
             [1, 2, 3], [2, 3, 4], ... etc
     """
 
-    def __init__(self,
-                 videofile: Union[str, os.PathLike],
-                 transform,
-                 sequence_length: int = 11,
-                 num_workers: int = 0,
-                 mean_by_channels: Union[list, np.ndarray] = [0, 0, 0]):
+    def __init__(
+        self,
+        videofile: Union[str, os.PathLike],
+        transform,
+        sequence_length: int = 11,
+        num_workers: int = 0,
+        mean_by_channels: Union[list, np.ndarray] = [0, 0, 0],
+    ):
         """Cosntructor for video iterable
 
         Parameters
@@ -98,10 +99,12 @@ class VideoIterable(data.IterableDataset):
         im = self.transform(im)
         self._image_shape = im.shape
 
-    def get_zeros_image(self,):
+    def get_zeros_image(
+        self,
+    ):
         if self._zeros_image is None:
             if self._image_shape is None:
-                raise ValueError('must set shape before getting zeros image')
+                raise ValueError("must set shape before getting zeros image")
             # ALWAYS ASSUME OUTPUT IS TRANSPOSED
             self._zeros_image = np.zeros(self._image_shape, dtype=np.uint8)
             for i in range(3):
@@ -115,12 +118,12 @@ class VideoIterable(data.IterableDataset):
             assert np.array_equal(np.clip(mean_by_channels, 0, 255), np.array(mean_by_channels))
             return np.array(mean_by_channels).astype(np.uint8)
         else:
-            raise ValueError('unexpected type for input channel mean: {}'.format(mean_by_channels))
+            raise ValueError("unexpected type for input channel mean: {}".format(mean_by_channels))
 
     def my_iter_func(self, start, end):
         for i in range(start, end):
             self.buffer.append(self.get_current_item())
-            yield {'images': np.stack(self.buffer, axis=1), 'framenum': self.cnt - 1 - self.sequence_length // 2}
+            yield {"images": np.stack(self.buffer, axis=1), "framenum": self.cnt - 1 - self.sequence_length // 2}
 
     def get_current_item(self):
         worker_info = data.get_worker_info()
@@ -134,8 +137,8 @@ class VideoIterable(data.IterableDataset):
         else:
             try:
                 im = self.readers[worker_id][self.cnt]
-            except Exception as e:
-                print(f'problem reading frame {self.cnt}')
+            except Exception:
+                print(f"problem reading frame {self.cnt}")
                 raise
             im = self.transform(im)
             # print(im.dtype)
@@ -184,10 +187,10 @@ class VideoIterable(data.IterableDataset):
                 continue
             try:
                 v.close()
-            except Exception as e:
-                print(f'error destroying reader {k}')
+            except Exception:
+                print(f"error destroying reader {k}")
             else:
-                print(f'destroyed {k}')
+                print(f"destroyed {k}")
 
     def __exit__(self, *args):
         self.close()
@@ -216,15 +219,17 @@ class SingleVideoDataset(data.Dataset):
         # ~5 x 11
     """
 
-    def __init__(self,
-                 videofile: Union[str, os.PathLike],
-                 labelfile: Union[str, os.PathLike] = None,
-                 mean_by_channels: Union[list, np.ndarray] = [0, 0, 0],
-                 frames_per_clip: int = 1,
-                 transform=None,
-                 reduce: bool = True,
-                 conv_mode: str = '2d',
-                 keep_reader_open: bool = False):
+    def __init__(
+        self,
+        videofile: Union[str, os.PathLike],
+        labelfile: Union[str, os.PathLike] = None,
+        mean_by_channels: Union[list, np.ndarray] = [0, 0, 0],
+        frames_per_clip: int = 1,
+        transform=None,
+        reduce: bool = True,
+        conv_mode: str = "2d",
+        keep_reader_open: bool = False,
+    ):
         """Initializes a VideoDataset object.
 
         Args:
@@ -252,33 +257,33 @@ class SingleVideoDataset(data.Dataset):
         self.supervised = self.labelfile is not None
 
         assert os.path.isfile(videofile) or os.path.isdir(videofile)
-        assert self.conv_mode in ['2d', '3d']
+        assert self.conv_mode in ["2d", "3d"]
 
         # find labels given the filename of a video, load, save as an attribute for fast reading
         if self.supervised:
             assert os.path.isfile(labelfile)
             # self.video_list, self.label_list = purge_unlabeled_videos(self.video_list, self.label_list)
-            labels, class_counts, num_labels, num_pos, num_neg = read_all_labels([self.labelfile],
-                                                                                 True,
-                                                                                 multilabel=not self.reduce)
+            labels, class_counts, num_labels, num_pos, num_neg = read_all_labels(
+                [self.labelfile], True, multilabel=not self.reduce
+            )
             self.labels = labels
             self.class_counts = class_counts
             self.num_labels = num_labels
             self.num_pos = num_pos
             self.num_neg = num_neg
-            log.debug('label shape: {}'.format(self.labels.shape))
+            log.debug("label shape: {}".format(self.labels.shape))
 
         metadata = {}
         ret, width, height, framecount = get_video_metadata(self.videofile)
         if ret:
-            metadata['name'] = videofile
-            metadata['width'] = width
-            metadata['height'] = height
-            metadata['framecount'] = framecount
+            metadata["name"] = videofile
+            metadata["width"] = width
+            metadata["height"] = height
+            metadata["framecount"] = framecount
         else:
-            raise ValueError('error loading video: {}'.format(videofile))
+            raise ValueError("error loading video: {}".format(videofile))
         self.metadata = metadata
-        self.N = self.metadata['framecount']
+        self.N = self.metadata["framecount"]
         self._zeros_image = None
 
     def get_zeros_image(self, c, h, w, channel_first: bool = True):
@@ -296,7 +301,7 @@ class SingleVideoDataset(data.Dataset):
             assert np.array_equal(np.clip(mean_by_channels, 0, 255), np.array(mean_by_channels))
             return np.array(mean_by_channels).astype(np.uint8)
         else:
-            raise ValueError('unexpected type for input channel mean: {}'.format(mean_by_channels))
+            raise ValueError("unexpected type for input channel mean: {}".format(mean_by_channels))
 
     def __len__(self):
         return self.N
@@ -329,7 +334,7 @@ class SingleVideoDataset(data.Dataset):
         # if frames per clip is 11, dataset[0] would have 5 blank frames preceding, with the 6th-11th being real frames
         blank_start_frames = max(self.frames_per_clip // 2 - index, 0)
 
-        framecount = self.metadata['framecount']
+        framecount = self.metadata["framecount"]
         # cap = cv2.VideoCapture(self.movies[style][movie_index])
         start_frame = index - self.frames_per_clip // 2 + blank_start_frames
         blank_end_frames = max(index - framecount + self.frames_per_clip // 2 + 1, 0)
@@ -342,8 +347,9 @@ class SingleVideoDataset(data.Dataset):
                     image = reader[i + start_frame]
                 except Exception as e:
                     image = self._zeros_image.copy().transpose(1, 2, 0)
-                    log.warning('Error {} on frame {} of video {}. Is the video corrupted?'.format(
-                        e, index, self.videofile))
+                    log.warning(
+                        "Error {} on frame {} of video {}. Is the video corrupted?".format(e, index, self.videofile)
+                    )
                 if self.transform:
                     random.seed(seed)
                     image = self.transform(image)
@@ -353,31 +359,34 @@ class SingleVideoDataset(data.Dataset):
         images = self.append_with_zeros(images, blank_end_frames)
 
         if log.isEnabledFor(logging.DEBUG):
-            log.debug('idx: {} st: {} blank_start: {} blank_end: {} real: {} total: {}'.format(
-                index, start_frame, blank_start_frames, blank_end_frames, real_frames, framecount))
+            log.debug(
+                "idx: {} st: {} blank_start: {} blank_end: {} real: {} total: {}".format(
+                    index, start_frame, blank_start_frames, blank_end_frames, real_frames, framecount
+                )
+            )
 
         # images are now numpy arrays of shape 3, H, W
         # stacking in the first dimension changes to 3, T, H, W, compatible with Conv3D
         images = np.stack(images, axis=1)
 
         if log.isEnabledFor(logging.DEBUG):
-            log.debug('images shape: {}'.format(images.shape))
+            log.debug("images shape: {}".format(images.shape))
         # print(images.shape)
-        outputs = {'images': images}
+        outputs = {"images": images}
         if self.supervised:
             label = self.labels[index]
             if self.reduce:
                 try:
                     label = np.where(label)[0][0].astype(np.int64)
                 except IndexError:
-                    logging.error(f'label {index} from video {self.videofile} has no positive labels! {label}')
+                    logging.error(f"label {index} from video {self.videofile} has no positive labels! {label}")
                     raise
-            outputs['labels'] = label
+            outputs["labels"] = label
         return outputs
 
 
 class VideoDataset(data.Dataset):
-    """ Simple wrapper around SingleVideoDataset for smoothly loading multiple videos """
+    """Simple wrapper around SingleVideoDataset for smoothly loading multiple videos"""
 
     def __init__(self, videofiles: list, labelfiles: list, *args, **kwargs):
         datasets, labels = [], []
@@ -418,33 +427,34 @@ class VideoDataset(data.Dataset):
 class SingleSequenceDataset(data.Dataset):
     """PyTorch Dataset for loading a set of saved 1d features and one-hot labels for Action Detection.
 
-        Features:
-            - Loads a set of sequential frames and sequential one-hot labels
-            - loads by indexing from an HDF5 dataset, given a dataset name (latent_name)
-            - Pads beginning or end so that every label has a corresponding clip
-            - Optionally loads two-stream features
+    Features:
+        - Loads a set of sequential frames and sequential one-hot labels
+        - loads by indexing from an HDF5 dataset, given a dataset name (latent_name)
+        - Pads beginning or end so that every label has a corresponding clip
+        - Optionally loads two-stream features
 
-        Example:
-            dataset = SequenceDataset(['features1.h5', 'features2.h5'], label_files=['labels1.csv', 'labels2.csv',
-                h5_key='CNN_features', sequence_length=180, is_two_stream=True)
-            features, labels = dataset(np.random.randint(low=0, high=len(dataset))
-            print(features.shape)
-            # 180 x 1024
-            print(labels.shape)
-            # assuming there are 5 classes in dataset
-            # ~5 x 180
-        """
+    Example:
+        dataset = SequenceDataset(['features1.h5', 'features2.h5'], label_files=['labels1.csv', 'labels2.csv',
+            h5_key='CNN_features', sequence_length=180, is_two_stream=True)
+        features, labels = dataset(np.random.randint(low=0, high=len(dataset))
+        print(features.shape)
+        # 180 x 1024
+        print(labels.shape)
+        # assuming there are 5 classes in dataset
+        # ~5 x 180
+    """
 
-    def __init__(self,
-                 data_file: Union[str, os.PathLike],
-                 labelfile: Union[str, os.PathLike],
-                 N: int,
-                 sequence_length: int = 60,
-                 nonoverlapping: bool = True,
-                 store_in_ram: bool = True,
-                 reduce: bool = False,
-                 stack_in_time: bool = False):
-
+    def __init__(
+        self,
+        data_file: Union[str, os.PathLike],
+        labelfile: Union[str, os.PathLike],
+        N: int,
+        sequence_length: int = 60,
+        nonoverlapping: bool = True,
+        store_in_ram: bool = True,
+        reduce: bool = False,
+        stack_in_time: bool = False,
+    ):
         self.reduce = reduce
 
         assert os.path.isfile(data_file)
@@ -475,7 +485,7 @@ class SingleSequenceDataset(data.Dataset):
         self.verify_dataset()
 
         tmp_sequence = self.__getitem__(0)  # self.read_sequence([0, 1])
-        self.num_features = tmp_sequence['features'].shape[0]
+        self.num_features = tmp_sequence["features"].shape[0]
 
     def read_sequence(self, indices):
         raise NotImplementedError
@@ -500,8 +510,7 @@ class SingleSequenceDataset(data.Dataset):
             inds = np.arange(self.N)
             self.starts = inds - self.sequence_length // 2
             # if it's odd, should go from
-            self.ends = inds + self.sequence_length//2 + \
-                self.sequence_length % 2
+            self.ends = inds + self.sequence_length // 2 + self.sequence_length % 2
 
     def __len__(self):
         return len(self.starts)
@@ -538,20 +547,24 @@ class SingleSequenceDataset(data.Dataset):
             label_indices = indices
             label_pad = pad
 
-        assert (len(indices) + pad_left + pad_right) == self.sequence_length, \
-                'indices: {} + pad_left: {} + pad_right: {} should equal seq len: {}'.format(
-                len(indices), pad_left, pad_right, self.sequence_length)
+        assert (len(indices) + pad_left + pad_right) == self.sequence_length, (
+            "indices: {} + pad_left: {} + pad_right: {} should equal seq len: {}".format(
+                len(indices), pad_left, pad_right, self.sequence_length
+            )
+        )
         # if we are stacking in time, label indices should not be the sequence length
         if not self.stack_in_time:
-            assert (len(label_indices) + label_pad[0] + label_pad[1]) == self.sequence_length, \
-                    'label indices: {} + pad_left: {} + pad_right: {} should equal seq len: {}'.format(
-                    len(label_indices), label_pad[0], label_pad[1], self.sequence_length)
+            assert (len(label_indices) + label_pad[0] + label_pad[1]) == self.sequence_length, (
+                "label indices: {} + pad_left: {} + pad_right: {} should equal seq len: {}".format(
+                    len(label_indices), label_pad[0], label_pad[1], self.sequence_length
+                )
+            )
         return indices, label_indices, pad, label_pad
 
     def __del__(self):
-        if hasattr(self, 'sequence'):
+        if hasattr(self, "sequence"):
             del self.sequence
-        if hasattr(self, 'labels'):
+        if hasattr(self, "labels"):
             del self.labels
 
     def __getitem__(self, index: int) -> dict:
@@ -565,7 +578,7 @@ class SingleSequenceDataset(data.Dataset):
         output = {}
         pad_left, pad_right = pad
         for key, value in data.items():
-            value = np.pad(value, ((0, 0), (pad_left, pad_right)), mode='constant')
+            value = np.pad(value, ((0, 0), (pad_left, pad_right)), mode="constant")
             if self.stack_in_time:
                 value = value.flatten()
             value = torch.from_numpy(value).float()
@@ -579,41 +592,43 @@ class SingleSequenceDataset(data.Dataset):
                 labels = self.label[:, label_indices].astype(np.int64)
                 if labels.ndim == 1:
                     labels = labels[:, np.newaxis]
-                labels = np.pad(labels, ((0, 0), (pad_left, pad_right)), mode='constant', constant_values=-1)
+                labels = np.pad(labels, ((0, 0), (pad_left, pad_right)), mode="constant", constant_values=-1)
             else:
                 labels = self.label[label_indices].astype(np.int64)
-                labels = np.pad(labels, (pad_left, pad_right), mode='constant', constant_values=-1)
+                labels = np.pad(labels, (pad_left, pad_right), mode="constant", constant_values=-1)
 
             # if we stack in time, we want to make sure we have labels of shape (N_behaviors,)
             # not (N_behaviors, 1)
             labels = labels.squeeze()
             labels = torch.from_numpy(labels).to(torch.long)
-            output['labels'] = labels
+            output["labels"] = labels
 
-            if labels.ndim > 1 and labels.shape[1] != output['features'].shape[1]:
-                out_shape = output['features'].shape
-                raise ValueError(f'problem in label shape! {labels.shape}, {out_shape}')
+            if labels.ndim > 1 and labels.shape[1] != output["features"].shape[1]:
+                out_shape = output["features"].shape
+                raise ValueError(f"problem in label shape! {labels.shape}, {out_shape}")
 
         return output
 
 
 class KeypointDataset(SingleSequenceDataset):
-    """Dataset for reading keypoints (e.g. from deeplabcut) and performing basis function expansion. 
-    
+    """Dataset for reading keypoints (e.g. from deeplabcut) and performing basis function expansion.
+
     Currently, only an edited variant of Sturman et al.'s basis expansion is implemented
-    Sturman, O., von Ziegler, L., Schläppi, C. et al. Deep learning-based behavioral analysis reaches human 
-        accuracy and is capable of outperforming commercial solutions. Neuropsychopharmacol. 45, 1942–1952 (2020). 
+    Sturman, O., von Ziegler, L., Schläppi, C. et al. Deep learning-based behavioral analysis reaches human
+        accuracy and is capable of outperforming commercial solutions. Neuropsychopharmacol. 45, 1942–1952 (2020).
         https://doi.org/10.1038/s41386-020-0776-y
     """
 
-    def __init__(self,
-                 data_file: Union[str, os.PathLike],
-                 labelfile: Union[str, os.PathLike],
-                 videofile: Union[str, os.PathLike],
-                 expansion_method: str = 'sturman',
-                 confidence_threshold: float = 0.9,
-                 *args,
-                 **kwargs):
+    def __init__(
+        self,
+        data_file: Union[str, os.PathLike],
+        labelfile: Union[str, os.PathLike],
+        videofile: Union[str, os.PathLike],
+        expansion_method: str = "sturman",
+        confidence_threshold: float = 0.9,
+        *args,
+        **kwargs,
+    ):
         """Constructor
 
         Parameters
@@ -632,10 +647,10 @@ class KeypointDataset(SingleSequenceDataset):
         Raises
         ------
         NotImplementedError
-            For basis function expansion. Currently, only 'sturman' is implemented: 
-            
+            For basis function expansion. Currently, only 'sturman' is implemented:
+
         """
-        if expansion_method == 'sturman':
+        if expansion_method == "sturman":
             self.expansion_func = expand_features_sturman
         else:
             raise NotImplementedError
@@ -672,45 +687,39 @@ class KeypointDataset(SingleSequenceDataset):
 
     def verify_dataset(self):
         if self.supervised:
-            assert self.label.shape[1] == self.sequence.shape[1], 'label {} and sequence {} shape do not match!'.format(
-                self.label.shape, self.sequence.shape)
+            assert self.label.shape[1] == self.sequence.shape[1], "label {} and sequence {} shape do not match!".format(
+                self.label.shape, self.sequence.shape
+            )
 
         assert self.sequence is not None
 
     def read_sequence(self, indices):
         data = {}
-        data['features'] = self.sequence[:, indices]
+        data["features"] = self.sequence[:, indices]
         return data
 
 
 class FeatureVectorDataset(SingleSequenceDataset):
-    """Reads image and flow feature vectors from HDF5 files. 
-    """
+    """Reads image and flow feature vectors from HDF5 files."""
 
-    def __init__(self,
-                 data_file,
-                 labelfile,
-                 h5_key: str,
-                 store_in_ram=False,
-                 is_two_stream: bool = True,
-                 *args,
-                 **kwargs):
-
+    def __init__(
+        self, data_file, labelfile, h5_key: str, store_in_ram=False, is_two_stream: bool = True, *args, **kwargs
+    ):
         self.is_two_stream = is_two_stream
         self.store_in_ram = store_in_ram
 
         assert os.path.isfile(data_file)
         self.key = h5_key
         if self.is_two_stream:
-            self.flow_key = self.key + '/flow_features'
-            self.image_key = self.key + '/spatial_features'
-        self.logit_key = self.key + '/logits'
+            self.flow_key = self.key + "/flow_features"
+            self.image_key = self.key + "/spatial_features"
+        self.logit_key = self.key + "/logits"
         self.data_file = data_file
 
         self.verify_dataset()
         data = self.read_features_from_disk(None, None)
 
-        features_shape = data['features'].shape
+        features_shape = data["features"].shape
         self.shape = features_shape
         self.N = self.shape[1]
         if self.store_in_ram:
@@ -722,7 +731,7 @@ class FeatureVectorDataset(SingleSequenceDataset):
         super().__init__(data_file, labelfile, self.N, *args, **kwargs)
 
     def verify_dataset(self):
-        with h5py.File(self.data_file, 'r') as f:
+        with h5py.File(self.data_file, "r") as f:
             assert self.logit_key in f
 
             if self.is_two_stream:
@@ -731,20 +740,17 @@ class FeatureVectorDataset(SingleSequenceDataset):
                 flow_shape = f[self.flow_key].shape
                 image_shape = f[self.image_key].shape
                 assert flow_shape[0] == image_shape[0]
-                # self.N = image_shape[0]
             else:
                 assert self.key in f
-                shape = f[self.key].shape
-                # self.N = shape[0]
 
     def read_features_from_disk(self, start_ind, end_ind):
         inds = slice(start_ind, end_ind)
-        with h5py.File(self.data_file, 'r') as f:
+        with h5py.File(self.data_file, "r") as f:
             if self.is_two_stream:
                 flow_shape = f[self.flow_key].shape
                 image_shape = f[self.image_key].shape
                 assert len(flow_shape) == 2
-                assert (flow_shape == image_shape)
+                assert flow_shape == image_shape
                 # we want each timepoint to be one COLUMN
                 flow_feats = f[self.flow_key][inds, :].T
                 image_feats = f[self.image_key][inds, :].T
@@ -757,7 +763,7 @@ class FeatureVectorDataset(SingleSequenceDataset):
 
     def read_sequence(self, indices):
         if self.store_in_ram:
-            data = {'features': self.data['features'][:, indices], 'logits': self.data['logits'][:, indices]}
+            data = {"features": self.data["features"][:, indices], "logits": self.data["logits"][:, indices]}
         else:
             # assume indices are in order
             # we use the start and end so that we can slice without knowing the exact size of the dataset
@@ -766,15 +772,11 @@ class FeatureVectorDataset(SingleSequenceDataset):
 
 
 class SequenceDataset(data.Dataset):
-    """ Simple wrapper around SingleSequenceDataset for smoothly loading multiple sequences """
+    """Simple wrapper around SingleSequenceDataset for smoothly loading multiple sequences"""
 
-    def __init__(self,
-                 datafiles: list,
-                 labelfiles: list,
-                 videofiles: list = None,
-                 is_keypoint: bool = False,
-                 *args,
-                 **kwargs):
+    def __init__(
+        self, datafiles: list, labelfiles: list, videofiles: list = None, is_keypoint: bool = False, *args, **kwargs
+    ):
         datasets = []
         for i, (datafile, labelfile) in enumerate(zip(datafiles, labelfiles)):
             if is_keypoint:
@@ -808,23 +810,25 @@ class SequenceDataset(data.Dataset):
         return self.dataset[index]
 
 
-def get_video_datasets(datadir: Union[str, os.PathLike],
-                       xform: dict,
-                       is_two_stream: bool = False,
-                       reload_split: bool = True,
-                       splitfile: Union[str, os.PathLike] = None,
-                       train_val_test: Union[list, np.ndarray] = [0.8, 0.1, 0.1],
-                       weight_exp: float = 1.0,
-                       rgb_frames: int = 1,
-                       flow_frames: int = 10,
-                       supervised=True,
-                       reduce=False,
-                       flow_max: int = 5,
-                       flow_style: str = 'linear',
-                       valid_splits_only: bool = True,
-                       conv_mode: str = '2d',
-                       mean_by_channels: list = [0.5, 0.5, 0.5]):
-    """ Gets dataloaders for video-based datasets.
+def get_video_datasets(
+    datadir: Union[str, os.PathLike],
+    xform: dict,
+    is_two_stream: bool = False,
+    reload_split: bool = True,
+    splitfile: Union[str, os.PathLike] = None,
+    train_val_test: Union[list, np.ndarray] = [0.8, 0.1, 0.1],
+    weight_exp: float = 1.0,
+    rgb_frames: int = 1,
+    flow_frames: int = 10,
+    supervised=True,
+    reduce=False,
+    flow_max: int = 5,
+    flow_style: str = "linear",
+    valid_splits_only: bool = True,
+    conv_mode: str = "2d",
+    mean_by_channels: list = [0.5, 0.5, 0.5],
+):
+    """Gets dataloaders for video-based datasets.
 
     Parameters
     ----------
@@ -882,11 +886,11 @@ def get_video_datasets(datadir: Union[str, os.PathLike],
         split contains the split dictionary, for saving
         keys for loss weighting are also added. see make_loss_weight for explanation
     """
-    return_types = ['rgb']
+    return_types = ["rgb"]
     if is_two_stream:
-        return_types += ['flow']
+        return_types += ["flow"]
     if supervised:
-        return_types += ['label']
+        return_types += ["label"]
     # records: dictionary of dictionaries. Keys: unique data identifiers
     # values: a dictionary corresponding to different files. the first record might be:
     # {'mouse000': {'rgb': path/to/rgb.avi, 'label':path/to/labels.csv} }
@@ -898,73 +902,77 @@ def get_video_datasets(datadir: Union[str, os.PathLike],
         records = purge_unlabeled_elements_from_records(records)
 
     if len(records) < 3:
-        error_message = 'You only have {} valid videos with file types {}!'.format(len(records), return_types)
-        error_message += 'You need at least 3 videos in your project to begin training.'
+        error_message = "You only have {} valid videos with file types {}!".format(len(records), return_types)
+        error_message += "You need at least 3 videos in your project to begin training."
         raise ValueError(error_message)
 
     # returns a dictionary, where each split in ['train', 'val', 'test'] as a list of keys
     # each key corresponds to a unique directory, and has
-    split_dictionary = get_split_from_records(records, datadir, splitfile, supervised, reload_split, valid_splits_only,
-                                              train_val_test)
+    split_dictionary = get_split_from_records(
+        records, datadir, splitfile, supervised, reload_split, valid_splits_only, train_val_test
+    )
     # it's possible that your split has records that are invalid for the current task.
     # e.g.: you've added a video, but not labeled it yet. In that case, it will already be in your split, but it is
     # invalid for current purposes, because it has no label. Therefore, we want to remove it from the current split
     split_dictionary = remove_invalid_records_from_split_dictionary(split_dictionary, records)
 
     datasets = {}
-    for i, split in enumerate(['train', 'val', 'test']):
-        rgb = [records[i]['rgb'] for i in split_dictionary[split]]
-        flow = [records[i]['flow'] for i in split_dictionary[split]]
+    for i, split in enumerate(["train", "val", "test"]):
+        rgb = [records[i]["rgb"] for i in split_dictionary[split]]
+        # flow = [records[i]["flow"] for i in split_dictionary[split]]
 
-        if split == 'test' and len(rgb) == 0:
+        if split == "test" and len(rgb) == 0:
             datasets[split] = None
             continue
 
         if supervised:
-            labelfiles = [records[i]['label'] for i in split_dictionary[split]]
+            labelfiles = [records[i]["label"] for i in split_dictionary[split]]
         else:
             labelfiles = None
 
-        datasets[split] = VideoDataset(rgb,
-                                       labelfiles,
-                                       frames_per_clip=rgb_frames,
-                                       reduce=reduce,
-                                       transform=xform[split],
-                                       conv_mode=conv_mode,
-                                       mean_by_channels=mean_by_channels)
-    data_info = {'split': split_dictionary}
+        datasets[split] = VideoDataset(
+            rgb,
+            labelfiles,
+            frames_per_clip=rgb_frames,
+            reduce=reduce,
+            transform=xform[split],
+            conv_mode=conv_mode,
+            mean_by_channels=mean_by_channels,
+        )
+    data_info = {"split": split_dictionary}
 
     if supervised:
-        data_info['class_counts'] = datasets['train'].class_counts
-        data_info['num_classes'] = len(data_info['class_counts'])
-        pos_weight, softmax_weight = make_loss_weight(data_info['class_counts'],
-                                                      datasets['train'].num_pos,
-                                                      datasets['train'].num_neg,
-                                                      weight_exp=weight_exp)
-        data_info['pos'] = datasets['train'].num_pos
-        data_info['neg'] = datasets['train'].num_neg
-        data_info['pos_weight'] = pos_weight
-        data_info['loss_weight'] = softmax_weight
+        data_info["class_counts"] = datasets["train"].class_counts
+        data_info["num_classes"] = len(data_info["class_counts"])
+        pos_weight, softmax_weight = make_loss_weight(
+            data_info["class_counts"], datasets["train"].num_pos, datasets["train"].num_neg, weight_exp=weight_exp
+        )
+        data_info["pos"] = datasets["train"].num_pos
+        data_info["neg"] = datasets["train"].num_neg
+        data_info["pos_weight"] = pos_weight
+        data_info["loss_weight"] = softmax_weight
 
     return datasets, data_info
 
 
-def get_sequence_datasets(datadir: Union[str, os.PathLike],
-                          latent_name: str,
-                          sequence_length: int = 60,
-                          is_two_stream: bool = True,
-                          nonoverlapping: bool = True,
-                          splitfile: str = None,
-                          reload_split: bool = True,
-                          store_in_ram: bool = False,
-                          train_val_test: Union[list, np.ndarray] = [0.8, 0.2, 0.0],
-                          weight_exp: float = 1.0,
-                          supervised=True,
-                          reduce=False,
-                          valid_splits_only: bool = True,
-                          is_keypoint: bool = False,
-                          stack_in_time: bool = False) -> Tuple[dict, dict]:
-    """ Gets dataloaders for sequence models assuming DeepEthogram file structure.
+def get_sequence_datasets(
+    datadir: Union[str, os.PathLike],
+    latent_name: str,
+    sequence_length: int = 60,
+    is_two_stream: bool = True,
+    nonoverlapping: bool = True,
+    splitfile: str = None,
+    reload_split: bool = True,
+    store_in_ram: bool = False,
+    train_val_test: Union[list, np.ndarray] = [0.8, 0.2, 0.0],
+    weight_exp: float = 1.0,
+    supervised=True,
+    reduce=False,
+    valid_splits_only: bool = True,
+    is_keypoint: bool = False,
+    stack_in_time: bool = False,
+) -> Tuple[dict, dict]:
+    """Gets dataloaders for sequence models assuming DeepEthogram file structure.
 
     Parameters
     ----------
@@ -1030,12 +1038,12 @@ def get_sequence_datasets(datadir: Union[str, os.PathLike],
     """
     return_types = []
     if is_keypoint:
-        log.info('Creating keypoint datasets, with feature expansion. Might take a few minutes')
-        return_types.append('keypoint')
+        log.info("Creating keypoint datasets, with feature expansion. Might take a few minutes")
+        return_types.append("keypoint")
     else:
-        return_types.append('output')
+        return_types.append("output")
     if supervised:
-        return_types.append('label')
+        return_types.append("label")
 
     # records: dictionary of dictionaries. Keys: unique data identifiers
     # values: a dictionary corresponding to different files. the first record might be:
@@ -1048,90 +1056,90 @@ def get_sequence_datasets(datadir: Union[str, os.PathLike],
         records = purge_unlabeled_elements_from_records(records)
 
     if len(records) < 3:
-        error_message = 'You only have {} valid videos with file types {}!'.format(len(records), return_types)
-        error_message += 'You need at least 3 videos in your project to begin training.'
+        error_message = "You only have {} valid videos with file types {}!".format(len(records), return_types)
+        error_message += "You need at least 3 videos in your project to begin training."
         raise ValueError(error_message)
 
     # returns a dictionary, where each split in ['train', 'val', 'test'] as a list of keys
     # each key corresponds to a unique directory, and has
-    split_dictionary = get_split_from_records(records, datadir, splitfile, supervised, reload_split, valid_splits_only,
-                                              train_val_test)
+    split_dictionary = get_split_from_records(
+        records, datadir, splitfile, supervised, reload_split, valid_splits_only, train_val_test
+    )
     # it's possible that your split has records that are invalid for the current task.
     # e.g.: you've added a video, but not labeled it yet. In that case, it will already be in your split, but it is
     # invalid for current purposes, because it has no label. Therefore, we want to remove it from the current split
     split_dictionary = remove_invalid_records_from_split_dictionary(split_dictionary, records)
-    # log.info('~~~~~ train val test split ~~~~~')
-    # pprint.pprint(split_dictionary)
 
-    splits = ['train', 'val', 'test']
+    splits = ["train", "val", "test"]
     datasets = {}
     # if stack_in_time, nonoverlapping would make us skip a bunch of labels
-    nonoverlapping = {'train': nonoverlapping, 'val': not stack_in_time, 'test': not stack_in_time}
+    nonoverlapping = {"train": nonoverlapping, "val": not stack_in_time, "test": not stack_in_time}
     for split in splits:
         if is_keypoint:
-            videofiles = [records[i]['rgb'] for i in split_dictionary[split]]
-            datafiles = [records[i]['keypoint'] for i in split_dictionary[split]]
+            videofiles = [records[i]["rgb"] for i in split_dictionary[split]]
+            datafiles = [records[i]["keypoint"] for i in split_dictionary[split]]
         else:
             videofiles = None
-            datafiles = [records[i]['output'] for i in split_dictionary[split]]
+            datafiles = [records[i]["output"] for i in split_dictionary[split]]
 
-        if split == 'test' and len(datafiles) == 0:
+        if split == "test" and len(datafiles) == 0:
             datasets[split] = None
             continue
-        # h5file, labelfile = outputs[i]
-        # print('making dataset:{}'.format(split))
 
         if supervised:
-            labelfiles = [records[i]['label'] for i in split_dictionary[split]]
+            labelfiles = [records[i]["label"] for i in split_dictionary[split]]
         else:
             labelfiles = None
 
         # todo: figure out a nice way to be able to pass arguments to one subclass that don't exist in the other
         # example: is_two_stream, latent_name
         if is_keypoint:
-            datasets[split] = SequenceDataset(datafiles,
-                                              labelfiles,
-                                              videofiles,
-                                              sequence_length=sequence_length,
-                                              nonoverlapping=nonoverlapping[split],
-                                              store_in_ram=store_in_ram,
-                                              reduce=reduce,
-                                              is_keypoint=is_keypoint,
-                                              stack_in_time=stack_in_time)
+            datasets[split] = SequenceDataset(
+                datafiles,
+                labelfiles,
+                videofiles,
+                sequence_length=sequence_length,
+                nonoverlapping=nonoverlapping[split],
+                store_in_ram=store_in_ram,
+                reduce=reduce,
+                is_keypoint=is_keypoint,
+                stack_in_time=stack_in_time,
+            )
         else:
-            datasets[split] = SequenceDataset(datafiles,
-                                              labelfiles,
-                                              videofiles=videofiles,
-                                              sequence_length=sequence_length,
-                                              h5_key=latent_name,
-                                              is_two_stream=is_two_stream,
-                                              nonoverlapping=nonoverlapping[split],
-                                              store_in_ram=store_in_ram,
-                                              reduce=reduce,
-                                              is_keypoint=is_keypoint,
-                                              stack_in_time=stack_in_time)
+            datasets[split] = SequenceDataset(
+                datafiles,
+                labelfiles,
+                videofiles=videofiles,
+                sequence_length=sequence_length,
+                h5_key=latent_name,
+                is_two_stream=is_two_stream,
+                nonoverlapping=nonoverlapping[split],
+                store_in_ram=store_in_ram,
+                reduce=reduce,
+                is_keypoint=is_keypoint,
+                stack_in_time=stack_in_time,
+            )
 
     # figure out what our inputs to our model will be (D dimension)
-    data_info = {'split': split_dictionary}
-    data_info['num_features'] = datasets['train'].num_features
+    data_info = {"split": split_dictionary}
+    data_info["num_features"] = datasets["train"].num_features
 
     if supervised:
-        data_info['class_counts'] = datasets['train'].class_counts
-        data_info['num_classes'] = len(data_info['class_counts'])
-        pos_weight, softmax_weight = make_loss_weight(data_info['class_counts'],
-                                                      datasets['train'].num_pos,
-                                                      datasets['train'].num_neg,
-                                                      weight_exp=weight_exp)
-        data_info['pos'] = datasets['train'].num_pos
-        data_info['neg'] = datasets['train'].num_neg
-        data_info['pos_weight'] = pos_weight
-        data_info['loss_weight'] = softmax_weight
+        data_info["class_counts"] = datasets["train"].class_counts
+        data_info["num_classes"] = len(data_info["class_counts"])
+        pos_weight, softmax_weight = make_loss_weight(
+            data_info["class_counts"], datasets["train"].num_pos, datasets["train"].num_neg, weight_exp=weight_exp
+        )
+        data_info["pos"] = datasets["train"].num_pos
+        data_info["neg"] = datasets["train"].num_neg
+        data_info["pos_weight"] = pos_weight
+        data_info["loss_weight"] = softmax_weight
 
     return datasets, data_info
 
 
 def get_datasets_from_cfg(cfg: DictConfig, model_type: str, input_images: int = 1) -> Tuple[dict, dict]:
-    """ Returns dataloader objects using a Hydra-generated configuration dictionary.
+    """Returns dataloader objects using a Hydra-generated configuration dictionary.
 
     This is the main entry point for getting dataloaders from the command line. it will return the correct dataloader
     with given hyperparameters for either flow, feature extractor, or sequence models.
@@ -1157,52 +1165,55 @@ def get_datasets_from_cfg(cfg: DictConfig, model_type: str, input_images: int = 
         information see the specific dataloader of the model you're training, e.g. get_video_dataloaders
     """
     #
-    supervised = model_type != 'flow_generator'
-    if model_type == 'feature_extractor' or model_type == 'flow_generator':
+    supervised = model_type != "flow_generator"
+    if model_type == "feature_extractor" or model_type == "flow_generator":
         arch = cfg[model_type].arch
-        mode = '3d' if '3d' in arch.lower() else '2d'
-        # log.info('getting dataloaders: {} convolution type detected'.format(mode))
+        mode = "3d" if "3d" in arch.lower() else "2d"
         xform = get_cpu_transforms(cfg.augs)
 
-        if cfg.project.name == 'kinetics':
+        if cfg.project.name == "kinetics":
             raise NotImplementedError
         else:
             reduce = False
-            if cfg.run.model == 'feature_extractor':
-                if cfg.feature_extractor.final_activation == 'softmax':
+            if cfg.run.model == "feature_extractor":
+                if cfg.feature_extractor.final_activation == "softmax":
                     reduce = True
-            datasets, info = get_video_datasets(datadir=cfg.project.data_path,
-                                                xform=xform,
-                                                is_two_stream=False,
-                                                reload_split=cfg.split.reload,
-                                                splitfile=cfg.split.file,
-                                                train_val_test=cfg.split.train_val_test,
-                                                weight_exp=cfg.train.loss_weight_exp,
-                                                rgb_frames=input_images,
-                                                supervised=supervised,
-                                                reduce=reduce,
-                                                valid_splits_only=True,
-                                                conv_mode=mode,
-                                                mean_by_channels=cfg.augs.normalization.mean)
+            datasets, info = get_video_datasets(
+                datadir=cfg.project.data_path,
+                xform=xform,
+                is_two_stream=False,
+                reload_split=cfg.split.reload,
+                splitfile=cfg.split.file,
+                train_val_test=cfg.split.train_val_test,
+                weight_exp=cfg.train.loss_weight_exp,
+                rgb_frames=input_images,
+                supervised=supervised,
+                reduce=reduce,
+                valid_splits_only=True,
+                conv_mode=mode,
+                mean_by_channels=cfg.augs.normalization.mean,
+            )
 
-    elif model_type == 'sequence':
-        if cfg.feature_extractor.final_activation == 'softmax':
+    elif model_type == "sequence":
+        if cfg.feature_extractor.final_activation == "softmax":
             reduce = True
-        datasets, info = get_sequence_datasets(cfg.project.data_path,
-                                               cfg.sequence.latent_name,
-                                               cfg.sequence.sequence_length,
-                                               is_two_stream=True,
-                                               nonoverlapping=cfg.sequence.nonoverlapping,
-                                               splitfile=cfg.split.file,
-                                               reload_split=True,
-                                               store_in_ram=False,
-                                               train_val_test=cfg.split.train_val_test,
-                                               weight_exp=cfg.train.loss_weight_exp,
-                                               supervised=True,
-                                               reduce=cfg.feature_extractor.final_activation == 'softmax',
-                                               valid_splits_only=True,
-                                               stack_in_time=cfg.sequence.arch == 'mlp',
-                                               is_keypoint=cfg.sequence.input_type == 'keypoints')
+        datasets, info = get_sequence_datasets(
+            cfg.project.data_path,
+            cfg.sequence.latent_name,
+            cfg.sequence.sequence_length,
+            is_two_stream=True,
+            nonoverlapping=cfg.sequence.nonoverlapping,
+            splitfile=cfg.split.file,
+            reload_split=True,
+            store_in_ram=False,
+            train_val_test=cfg.split.train_val_test,
+            weight_exp=cfg.train.loss_weight_exp,
+            supervised=True,
+            reduce=cfg.feature_extractor.final_activation == "softmax",
+            valid_splits_only=True,
+            stack_in_time=cfg.sequence.arch == "mlp",
+            is_keypoint=cfg.sequence.input_type == "keypoints",
+        )
     else:
-        raise ValueError('Unknown model type: {}'.format(model_type))
+        raise ValueError("Unknown model type: {}".format(model_type))
     return datasets, info
