@@ -12,6 +12,7 @@ from tqdm import tqdm
 from vidio import VideoReader
 
 from deepethogram import utils
+
 # from deepethogram.dataloaders import log
 from deepethogram.file_io import read_labels
 
@@ -26,8 +27,8 @@ def purge_unlabeled_videos(video_list: list, label_list: list) -> Tuple[list, li
     valid_videos = []
     valid_labels = []
 
-    warning_string = '''Labelfile {} associated with video {} has unlabeled frames! 
-        Please finish labeling or click the Finalize Labels button on the GUI.'''
+    warning_string = """Labelfile {} associated with video {} has unlabeled frames!
+        Please finish labeling or click the Finalize Labels button on the GUI."""
 
     for i in range(len(label_list)):
         label = read_labels(label_list[i])
@@ -43,16 +44,18 @@ def purge_unlabeled_videos(video_list: list, label_list: list) -> Tuple[list, li
 def purge_unlabeled_elements_from_records(records: dict) -> dict:
     valid_records = {}
 
-    warning_message = '''labelfile {} has unlabeled frames! 
+    warning_message = """labelfile {} has unlabeled frames!
         Please finish labeling or click the Finalize Labels button on the GUI.
-        Associated files: {}'''
+        Associated files: {}"""
 
     for animal, record in records.items():
-        labelfile = record['label']
+        labelfile = record["label"]
 
         if labelfile is None:
-            log.warning('Record {} does not have a labelfile! Please start and finish labeling. '.format(animal) + \
-                'Associated files: {}'.format(record))
+            log.warning(
+                "Record {} does not have a labelfile! Please start and finish labeling. ".format(animal)
+                + "Associated files: {}".format(record)
+            )
             continue
         label = read_labels(labelfile)
         has_unlabeled_frames = np.any(label == -1)
@@ -63,11 +66,10 @@ def purge_unlabeled_elements_from_records(records: dict) -> dict:
     return valid_records
 
 
-def make_loss_weight(class_counts: np.ndarray,
-                     num_pos: np.ndarray,
-                     num_neg: np.ndarray,
-                     weight_exp: float = 1) -> Tuple[np.ndarray, np.ndarray]:
-    """ Makes weight for different classes in loss function.
+def make_loss_weight(
+    class_counts: np.ndarray, num_pos: np.ndarray, num_neg: np.ndarray, weight_exp: float = 1
+) -> Tuple[np.ndarray, np.ndarray]:
+    """Makes weight for different classes in loss function.
 
     In general, rare classes will be up-weighted and common classes will be down-weighted.
 
@@ -95,7 +97,7 @@ def make_loss_weight(class_counts: np.ndarray,
     # if there are zero positive examples, we don't want the pos weight to be a large number
     # we want it to be infinity, then we will manually set it to zero
     with warnings.catch_warnings():
-        warnings.filterwarnings('ignore', category=RuntimeWarning)
+        warnings.filterwarnings("ignore", category=RuntimeWarning)
         pos_weight = num_neg / num_pos
     # if there are zero negative examples, loss should be 1
     pos_weight[pos_weight == 0] = 1
@@ -105,24 +107,24 @@ def make_loss_weight(class_counts: np.ndarray,
     pos_weight_transformed = np.nan_to_num(pos_weight_transformed, nan=0.0, posinf=0.0, neginf=0)
 
     with warnings.catch_warnings():
-        warnings.filterwarnings('ignore', category=RuntimeWarning)
+        warnings.filterwarnings("ignore", category=RuntimeWarning)
         softmax_weight = 1 / class_counts
     softmax_weight = np.nan_to_num(softmax_weight, nan=0.0, posinf=0.0, neginf=0)
     softmax_weight = softmax_weight / np.sum(softmax_weight)
     softmax_weight_transformed = (softmax_weight**weight_exp).astype(np.float32)
 
     np.set_printoptions(suppress=True)
-    log.info('Class counts: {}'.format(class_counts))
-    log.info('Pos weight: {}'.format(pos_weight))
-    log.info('Pos weight, weighted: {}'.format(pos_weight_transformed))
-    log.info('Softmax weight: {}'.format(softmax_weight))
-    log.info('Softmax weight transformed: {}'.format(softmax_weight_transformed))
+    log.info("Class counts: {}".format(class_counts))
+    log.info("Pos weight: {}".format(pos_weight))
+    log.info("Pos weight, weighted: {}".format(pos_weight_transformed))
+    log.info("Softmax weight: {}".format(softmax_weight))
+    log.info("Softmax weight transformed: {}".format(softmax_weight_transformed))
 
     return pos_weight_transformed, softmax_weight_transformed
 
 
 def get_video_metadata(videofile):
-    """ Simple wrapper to get video availability, width, height, and frame number """
+    """Simple wrapper to get video availability, width, height, and frame number"""
     try:
         with VideoReader(videofile) as reader:
             framenum = reader.nframes
@@ -133,24 +135,23 @@ def get_video_metadata(videofile):
     except BaseException as e:
         ret = False
         print(e)
-        print('Error reading file {}'.format(videofile))
+        print("Error reading file {}".format(videofile))
     return ret, width, height, framenum
 
 
 def extract_metadata(splitdir, allmovies=None, is_flow=False, num_workers=32):
-    """ Function to get the video metadata for all videos in Kinetics """
+    """Function to get the video metadata for all videos in Kinetics"""
     actions = os.listdir(splitdir)
     actions.sort()
 
     if allmovies is None:
-        allmovies = glob.glob(splitdir + '**/**/**.mp4') + glob.glob(splitdir + '**/**/**.avi')
+        allmovies = glob.glob(splitdir + "**/**/**.mp4") + glob.glob(splitdir + "**/**/**.avi")
     allmovies.sort()
 
     if not is_flow:
-        allmovies = [i for i in allmovies if 'flow' not in os.path.basename(i)]
+        allmovies = [i for i in allmovies if "flow" not in os.path.basename(i)]
     else:
-
-        allmovies = [i for i in allmovies if 'flow' in os.path.basename(i)]
+        allmovies = [i for i in allmovies if "flow" in os.path.basename(i)]
     widths = []
     heights = []
     framenums = []
@@ -166,9 +167,9 @@ def extract_metadata(splitdir, allmovies=None, is_flow=False, num_workers=32):
             # movies = glob.glob(action_dir + '**/**.mp4') + glob.glob(action_dir + '**/**.avi')
             movies.sort()
             if not is_flow:
-                movies = [i for i in movies if 'flow' not in os.path.basename(i)]
+                movies = [i for i in movies if "flow" not in os.path.basename(i)]
             else:
-                movies = [i for i in movies if 'flow' in os.path.basename(i)]
+                movies = [i for i in movies if "flow" in os.path.basename(i)]
             results = pool.map(get_video_metadata, movies)
 
             success = []
@@ -189,64 +190,62 @@ def extract_metadata(splitdir, allmovies=None, is_flow=False, num_workers=32):
                     action_indices.append(action_index)
 
     video_data = {
-        'name': allnames,
-        'action': allactions,
-        'action_int': action_indices,
-        'width': widths,
-        'height': heights,
-        'framecount': framenums
+        "name": allnames,
+        "action": allactions,
+        "action_int": action_indices,
+        "width": widths,
+        "height": heights,
+        "framecount": framenums,
     }
     df = pd.DataFrame(data=video_data)
-    fname = '_metadata.csv'
+    fname = "_metadata.csv"
     if is_flow:
-        fname = '_flow' + fname
+        fname = "_flow" + fname
     df.to_csv(os.path.join(os.path.dirname(splitdir), os.path.basename(splitdir) + fname))
     return df
 
 
 def find_labelfile(video: Union[str, os.PathLike]) -> Tuple[str, str]:
-    """ Function for finding a label file for a given a video """
+    """Function for finding a label file for a given a video"""
     base = os.path.splitext(video)[0]
-    labelfile = base + '_labels.csv'
+    labelfile = base + "_labels.csv"
     if os.path.isfile(labelfile):
-        return (labelfile, 'csv')
-    labelfile = base + '_labels.h5'
+        return (labelfile, "csv")
+    labelfile = base + "_labels.h5"
     if os.path.isfile(labelfile):
-        return (labelfile, 'h5')
-    labelfile = base + '_scores.csv'
+        return (labelfile, "h5")
+    labelfile = base + "_scores.csv"
     if os.path.isfile(labelfile):
-        return (labelfile, 'csv')
-    labelfile = base + '_scores.h5'
+        return (labelfile, "csv")
+    labelfile = base + "_scores.h5"
     if os.path.isfile(labelfile):
-        return (labelfile, 'h5')
+        return (labelfile, "h5")
     basedir = os.path.dirname(video)
     files = os.listdir(basedir)
     files.sort()
     files = [os.path.join(basedir, i) for i in files]
     # handles case where directory contains 'movie.avi', and 'labels.csv'
-    files = [i for i in files if 'label' in i or 'score' in i]
+    files = [i for i in files if "label" in i or "score" in i]
     if len(files) == 1:
-        if files[0].endswith('csv'):
-            return files[0], 'csv'
-        elif files[0].endswith('h5'):
-            return files[0], 'h5'
-    basename = os.path.basename(base).split('_')[:-1]
-    basename = '_'.join(basename)
+        if files[0].endswith("csv"):
+            return files[0], "csv"
+        elif files[0].endswith("h5"):
+            return files[0], "h5"
+    basename = os.path.basename(base).split("_")[:-1]
+    basename = "_".join(basename)
     matching_files = [i for i in files if basename in i]
     if len(matching_files) == 1:
         labelfile = matching_files[0]
         ext = os.path.splitext(labelfile)[1][1:]
         return labelfile, ext
-    raise ValueError('no corresponding labels found: {}'.format(video))
+    raise ValueError("no corresponding labels found: {}".format(video))
 
 
 def read_all_labels(labelfiles: list, fix: bool = True, multilabel: bool = True):
-    """ Function for reading all labels into memory """
+    """Function for reading all labels into memory"""
     labels = []
     for i, labelfile in enumerate(labelfiles):
-        assert (os.path.isfile(labelfile))
-        label_type = os.path.splitext(labelfile)[1][1:]
-        # labelfile, label_type = find_labelfile(video)
+        assert os.path.isfile(labelfile)
         label = read_labels(labelfile)
         H, W = label.shape
         # labels should be time x num_behaviors
@@ -254,7 +253,7 @@ def read_all_labels(labelfiles: list, fix: bool = True, multilabel: bool = True)
             label = label.T
         if label.shape[1] == 1:
             # add a background class
-            warnings.warn('binary labels found, adding background class')
+            warnings.warn("binary labels found, adding background class")
             label = np.hstack((np.logical_not(label), label))
 
         if fix:
@@ -307,12 +306,12 @@ def parse_split(split: Union[tuple, list, np.ndarray], N: int):
         N = total
         split = split / split.sum()
     else:
-        raise ValueError('Unknown split type: {}'.format(split.dtype))
+        raise ValueError("Unknown split type: {}".format(split.dtype))
     return split, N
 
 
 def train_val_test_split(records: dict, split: Union[tuple, list, np.ndarray] = (0.7, 0.15, 0.15)) -> dict:
-    """ Split a dict of dicts into train, validation, and test sets.
+    """Split a dict of dicts into train, validation, and test sets.
 
     Parameters
     ----------
@@ -340,28 +339,28 @@ def train_val_test_split(records: dict, split: Union[tuple, list, np.ndarray] = 
 
     # in place
     indices = np.random.permutation(N)
-    splits = ['train', 'val', 'test']
+    splits = ["train", "val", "test"]
     keys = np.array(keys)
     outputs = {}
-    outputs['metadata'] = {'split': split.tolist()}
+    outputs["metadata"] = {"split": split.tolist()}
     # print(list(split))
     # outputs['metadata']['split'] = split.tolist()
     # handle edge cases
     if len(records) < 4:
         assert len(records) > 1
-        warnings.warn('Only {} records found...'.format(len(keys)))
+        warnings.warn("Only {} records found...".format(len(keys)))
         shuffled = np.random.permutation(keys)
-        outputs['train'] = [str(shuffled[0])]
-        outputs['val'] = [str(shuffled[1])]
-        outputs['test'] = []
+        outputs["train"] = [str(shuffled[0])]
+        outputs["val"] = [str(shuffled[1])]
+        outputs["test"] = []
         if len(records) == 3:
             shuffled = np.random.permutation(keys)
-            outputs['test'] = [str(shuffled[2])]
+            outputs["test"] = [str(shuffled[2])]
         return outputs
 
     for i, spl in enumerate(splits):
         shuffled = keys[indices]
-        splitfiles = shuffled[starts[i]:ends[i]]
+        splitfiles = shuffled[starts[i] : ends[i]]
         outputs[spl] = splitfiles.tolist()
 
     # print(type(split.tolist()[0]))
@@ -369,21 +368,21 @@ def train_val_test_split(records: dict, split: Union[tuple, list, np.ndarray] = 
 
 
 def do_all_classes_have_labels(records: dict, split_dict: dict) -> bool:
-    """ Helper function to determine if each split has at least one instance of every class """
+    """Helper function to determine if each split has at least one instance of every class"""
     labelfiles = []
 
-    for split in ['train', 'val', 'test']:
+    for split in ["train", "val", "test"]:
         if len(split_dict[split]) > 0:
             splitfiles = split_dict[split]
             for f in splitfiles:
-                labelfiles.append(records[f]['label'])
+                labelfiles.append(records[f]["label"])
             # labelfiles += [records[i]['label'] for i in split_dict[split]]
     _, class_counts, _, _, _ = read_all_labels(labelfiles)
     return np.all(class_counts > 0)
 
 
 def get_valid_split(records: dict, train_val_test: Union[list, np.ndarray]) -> dict:
-    """  Gets a train, val, test split with at least one instance of every class
+    """Gets a train, val, test split with at least one instance of every class
 
     Keep doing train_test_split until each split of the data has at least one single example of every behavior
     in the dataset. it would be bad if your train data had class counts: [1000, 0, 0, 10] and your test data had
@@ -409,11 +408,11 @@ def get_valid_split(records: dict, train_val_test: Union[list, np.ndarray]) -> d
         split_dict = train_val_test_split(records, train_val_test)
         should_continue = do_all_classes_have_labels(records, split_dict)
         if not should_continue:
-            warnings.warn('Not all classes in the dataset have *any* labels!')
+            warnings.warn("Not all classes in the dataset have *any* labels!")
             return split_dict
         is_wrong = False
-        for split in ['train', 'val', 'test']:
-            labelfiles = [records[i]['label'] for i in split_dict[split]]
+        for split in ["train", "val", "test"]:
+            labelfiles = [records[i]["label"] for i in split_dict[split]]
             if len(labelfiles) > 0:
                 _, class_counts, _, _, _ = read_all_labels(labelfiles)
                 if not np.all(class_counts > 0):
@@ -422,12 +421,12 @@ def get_valid_split(records: dict, train_val_test: Union[list, np.ndarray]) -> d
 
 
 def update_split(records: dict, split_dictionary: dict) -> dict:
-    """ Updates existing split if there are new entries in the records dictionary """
+    """Updates existing split if there are new entries in the records dictionary"""
     # records: dictionary of dictionaries. Keys: unique data identifiers
     # values: a dictionary corresponding to different files. the first record might be:
     # {'mouse000': {'rgb': path/to/rgb.avi, 'label':path/to/labels.csv} }
     # split_dictionary: {'metadata': ..., 'train':[mouse000, mouse001], 'val':[mouse002,mouse003]... etc}
-    old_dictionary = {k: v for (k, v) in split_dictionary.items() if k != 'metadata'}
+    old_dictionary = {k: v for (k, v) in split_dictionary.items() if k != "metadata"}
     # https://stackoverflow.com/questions/952914/how-to-make-a-flat-list-out-of-list-of-lists
     old_keys = [item for sublist in old_dictionary.values() for item in sublist]
     old_keys.sort()
@@ -437,29 +436,31 @@ def update_split(records: dict, split_dictionary: dict) -> dict:
     # of the split dictionary
     new_entries = [i for i in new_keys if i not in old_keys]
     splits = list(split_dictionary.keys())
-    splits = [i for i in splits if i != 'metadata']
+    splits = [i for i in splits if i != "metadata"]
     if len(splits) == 3:
         # alphabetical order does not work
-        splits = ['train', 'val', 'test']
+        splits = ["train", "val", "test"]
     # goes through new entries, and assigns them to a split based on loaded split_probabilities
     if len(new_entries) > 0:
-        split_p = split_dictionary['metadata']['split']
+        split_p = split_dictionary["metadata"]["split"]
         N = len(new_entries)
         new_splits = np.random.choice(splits, size=(N,), p=split_p).tolist()
         for i, k in enumerate(new_entries):
             split_dictionary[new_splits[i]].append(k)
-            log.info('file {} assigned to split {}'.format(k, new_splits[i]))
+            log.info("file {} assigned to split {}".format(k, new_splits[i]))
     return split_dictionary
 
 
-def get_split_from_records(records: dict,
-                           datadir: Union[str, bytes, os.PathLike],
-                           splitfile: Union[str, bytes, os.PathLike] = None,
-                           supervised: bool = True,
-                           reload_split: bool = True,
-                           valid_splits_only: bool = True,
-                           train_val_test: list = [0.7, 0.15, 0.15]):
-    """ Splits the records into train, validation, and test splits
+def get_split_from_records(
+    records: dict,
+    datadir: Union[str, bytes, os.PathLike],
+    splitfile: Union[str, bytes, os.PathLike] = None,
+    supervised: bool = True,
+    reload_split: bool = True,
+    valid_splits_only: bool = True,
+    train_val_test: list = [0.7, 0.15, 0.15],
+):
+    """Splits the records into train, validation, and test splits
 
     Parameters
     ----------
@@ -484,9 +485,9 @@ def get_split_from_records(records: dict,
         see train_val_test_split
     """
     if splitfile is None:
-        splitfile = os.path.join(datadir, 'split.yaml')
+        splitfile = os.path.join(datadir, "split.yaml")
     else:
-        assert os.path.isfile(splitfile), 'split file does not exist! {}'.format(splitfile)
+        assert os.path.isfile(splitfile), "split file does not exist! {}".format(splitfile)
 
     if supervised and valid_splits_only:
         # this function makes sure that each split has all classes in the dataset
@@ -510,13 +511,13 @@ def get_split_from_records(records: dict,
 
 
 def remove_invalid_records_from_split_dictionary(split_dictionary: dict, records: dict) -> dict:
-    """ Removes records that exist in split_dictionary but not in records.
+    """Removes records that exist in split_dictionary but not in records.
     Can be useful if you previously had a video in your project and used that to make a train / val / test split,
     but later deleted it.
     """
     valid_records = {}
     record_keys = list(records.keys())
-    for split in ['train', 'val', 'test']:
+    for split in ["train", "val", "test"]:
         valid_records[split] = {}
         splitfiles = split_dictionary[split]
         for i, key in enumerate(record_keys):
@@ -536,7 +537,7 @@ def count_multilabeled_frames(label):
 def fix_label(labelfile, label: np.ndarray, multilabel: bool = True) -> np.ndarray:
     n_unlabeled = count_unlabeled_frames(label)
     if n_unlabeled > 0:
-        logging.warning(f'file {labelfile} has {n_unlabeled} unlabeled frames!! setting to background...')
+        logging.warning(f"file {labelfile} has {n_unlabeled} unlabeled frames!! setting to background...")
         # set rows to 0
         unlabeled = label.sum(axis=1) < 1
         label[unlabeled, :] = 0
@@ -547,7 +548,7 @@ def fix_label(labelfile, label: np.ndarray, multilabel: bool = True) -> np.ndarr
         # labels must be mutually exclusive
         n_multilabel = count_multilabeled_frames(label)
         if n_multilabel > 1:
-            logging.warning(f'file {labelfile} has {n_multilabel} multi-label frames! randomly selecting the label...')
+            logging.warning(f"file {labelfile} has {n_multilabel} multi-label frames! randomly selecting the label...")
             multilabel = label.sum(axis=1) > 1
             multilabel_inds = np.where(multilabel)[0]
             for ind in multilabel_inds:
