@@ -37,13 +37,14 @@ import torch.nn.functional as F
 import torch.utils.model_zoo as model_zoo
 
 
-__all__ = ['Inception3', 'inception_v3']
+__all__ = ["Inception3", "inception_v3"]
 
 
 model_urls = {
     # Inception v3 ported from TensorFlow
-    'inception_v3_google': 'https://download.pytorch.org/models/inception_v3_google-1a9a5a14.pth',
+    "inception_v3_google": "https://download.pytorch.org/models/inception_v3_google-1a9a5a14.pth",
 }
+
 
 def inception_v3(pretrained=False, in_channels=3, **kwargs):
     r"""Inception v3 model architecture from
@@ -53,15 +54,15 @@ def inception_v3(pretrained=False, in_channels=3, **kwargs):
         pretrained (bool): If True, returns a model pre-trained on ImageNet
     """
     if pretrained:
-        if 'transform_input' not in kwargs:
-            kwargs['transform_input'] = True
-        model = Inception3(in_channels=3,**kwargs)
+        if "transform_input" not in kwargs:
+            kwargs["transform_input"] = True
+        model = Inception3(in_channels=3, **kwargs)
         # from Wang et al. 2015: Towards good practices for very deep two-stream convnets
-        state_dict = model_zoo.load_url(model_zoo.load_url(model_urls['inception_v3_google']))
-        if in_channels !=3:
+        state_dict = model_zoo.load_url(model_zoo.load_url(model_urls["inception_v3_google"]))
+        if in_channels != 3:
             rgb_kernel_key = list(state_dict.keys())[0]
             rgb_kernel = state_dict[rgb_kernel_key]
-            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels,1,1)
+            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels, 1, 1)
             state_dict[rgb_kernel_key] = flow_kernel
             state_dict.update(state_dict)
         model.load_state_dict(state_dict)
@@ -69,9 +70,9 @@ def inception_v3(pretrained=False, in_channels=3, **kwargs):
 
     return Inception3(**kwargs)
 
-class Inception3(nn.Module):
 
-    def __init__(self, in_channels=3,num_classes=1000, aux_logits=True, transform_input=False, dropout_p=0):
+class Inception3(nn.Module):
+    def __init__(self, in_channels=3, num_classes=1000, aux_logits=True, transform_input=False, dropout_p=0):
         super(Inception3, self).__init__()
         self.aux_logits = aux_logits
         self.transform_input = transform_input
@@ -93,7 +94,7 @@ class Inception3(nn.Module):
         self.Mixed_7a = InceptionD(768)
         self.Mixed_7b = InceptionE(1280)
         self.Mixed_7c = InceptionE(2048)
-        self.avgpool = nn.AdaptiveAvgPool2d((1,1))
+        self.avgpool = nn.AdaptiveAvgPool2d((1, 1))
 
         self.dropout_p = dropout_p
         if self.dropout_p > 0:
@@ -104,7 +105,8 @@ class Inception3(nn.Module):
         for m in self.modules():
             if isinstance(m, nn.Conv2d) or isinstance(m, nn.Linear):
                 import scipy.stats as stats
-                stddev = m.stddev if hasattr(m, 'stddev') else 0.1
+
+                stddev = m.stddev if hasattr(m, "stddev") else 0.1
                 X = stats.truncnorm(-2, 2, scale=stddev)
                 values = torch.Tensor(X.rvs(m.weight.numel()))
                 values = values.view(m.weight.size())
@@ -178,7 +180,6 @@ class Inception3(nn.Module):
 
 
 class InceptionA(nn.Module):
-
     def __init__(self, in_channels, pool_features):
         super(InceptionA, self).__init__()
         self.branch1x1 = BasicConv2d(in_channels, 64, kernel_size=1)
@@ -210,7 +211,6 @@ class InceptionA(nn.Module):
 
 
 class InceptionB(nn.Module):
-
     def __init__(self, in_channels):
         super(InceptionB, self).__init__()
         self.branch3x3 = BasicConv2d(in_channels, 384, kernel_size=3, stride=2)
@@ -233,7 +233,6 @@ class InceptionB(nn.Module):
 
 
 class InceptionC(nn.Module):
-
     def __init__(self, in_channels, channels_7x7):
         super(InceptionC, self).__init__()
         self.branch1x1 = BasicConv2d(in_channels, 192, kernel_size=1)
@@ -272,7 +271,6 @@ class InceptionC(nn.Module):
 
 
 class InceptionD(nn.Module):
-
     def __init__(self, in_channels):
         super(InceptionD, self).__init__()
         self.branch3x3_1 = BasicConv2d(in_channels, 192, kernel_size=1)
@@ -298,7 +296,6 @@ class InceptionD(nn.Module):
 
 
 class InceptionE(nn.Module):
-
     def __init__(self, in_channels):
         super(InceptionE, self).__init__()
         self.branch1x1 = BasicConv2d(in_channels, 320, kernel_size=1)
@@ -340,7 +337,6 @@ class InceptionE(nn.Module):
 
 
 class InceptionAux(nn.Module):
-
     def __init__(self, in_channels, num_classes):
         super(InceptionAux, self).__init__()
         self.conv0 = BasicConv2d(in_channels, 128, kernel_size=1)
@@ -365,7 +361,6 @@ class InceptionAux(nn.Module):
 
 
 class BasicConv2d(nn.Module):
-
     def __init__(self, in_channels, out_channels, **kwargs):
         super(BasicConv2d, self).__init__()
         self.conv = nn.Conv2d(in_channels, out_channels, bias=False, **kwargs)
