@@ -30,7 +30,6 @@ from deepethogram.utils import load_state
 import warnings
 
 
-
 # __all__ = [
 #     'ResNet', 'resnet10_3d', 'resnet18_3d', 'resnet34_3d', 'resnet50_3d', 'resnet101_3d',
 #     'resnet152_3d', 'resnet200_3d'
@@ -39,20 +38,12 @@ import warnings
 
 def conv3x3x3(in_planes, out_planes, stride=1):
     # 3x3x3 convolution with padding
-    return nn.Conv3d(
-        in_planes,
-        out_planes,
-        kernel_size=3,
-        stride=stride,
-        padding=1,
-        bias=False)
+    return nn.Conv3d(in_planes, out_planes, kernel_size=3, stride=stride, padding=1, bias=False)
 
 
 def downsample_basic_block(x, planes, stride):
     out = F.avg_pool3d(x, kernel_size=1, stride=stride)
-    zero_pads = torch.Tensor(
-        out.size(0), planes - out.size(1), out.size(2), out.size(3),
-        out.size(4)).zero_()
+    zero_pads = torch.Tensor(out.size(0), planes - out.size(1), out.size(2), out.size(3), out.size(4)).zero_()
     if isinstance(out.data, torch.cuda.FloatTensor):
         zero_pads = zero_pads.cuda()
 
@@ -100,8 +91,7 @@ class Bottleneck(nn.Module):
         super(Bottleneck, self).__init__()
         self.conv1 = nn.Conv3d(inplanes, planes, kernel_size=1, bias=False)
         self.bn1 = nn.BatchNorm3d(planes)
-        self.conv2 = nn.Conv3d(
-            planes, planes, kernel_size=3, stride=stride, padding=1, bias=False)
+        self.conv2 = nn.Conv3d(planes, planes, kernel_size=3, stride=stride, padding=1, bias=False)
         self.bn2 = nn.BatchNorm3d(planes)
         self.conv3 = nn.Conv3d(planes, planes * 4, kernel_size=1, bias=False)
         self.bn3 = nn.BatchNorm3d(planes * 4)
@@ -133,33 +123,17 @@ class Bottleneck(nn.Module):
 
 
 class ResNet(nn.Module):
-
-    def __init__(self,
-                 block,
-                 layers,
-                 in_channels=3, 
-                 shortcut_type='B',
-                 num_classes=400,
-                dropout_p=0.5):
+    def __init__(self, block, layers, in_channels=3, shortcut_type="B", num_classes=400, dropout_p=0.5):
         self.inplanes = 64
         super(ResNet, self).__init__()
-        self.conv1 = nn.Conv3d(
-            in_channels,
-            64,
-            kernel_size=7,
-            stride=(1, 2, 2),
-            padding=(3, 3, 3),
-            bias=False)
+        self.conv1 = nn.Conv3d(in_channels, 64, kernel_size=7, stride=(1, 2, 2), padding=(3, 3, 3), bias=False)
         self.bn1 = nn.BatchNorm3d(64)
         self.relu = nn.ReLU(inplace=True)
         self.maxpool = nn.MaxPool3d(kernel_size=(3, 3, 3), stride=2, padding=1)
         self.layer1 = self._make_layer(block, 64, layers[0], shortcut_type)
-        self.layer2 = self._make_layer(
-            block, 128, layers[1], shortcut_type, stride=2)
-        self.layer3 = self._make_layer(
-            block, 256, layers[2], shortcut_type, stride=2)
-        self.layer4 = self._make_layer(
-            block, 512, layers[3], shortcut_type, stride=2)
+        self.layer2 = self._make_layer(block, 128, layers[1], shortcut_type, stride=2)
+        self.layer3 = self._make_layer(block, 256, layers[2], shortcut_type, stride=2)
+        self.layer4 = self._make_layer(block, 512, layers[3], shortcut_type, stride=2)
         #         last_duration = int(math.ceil(sample_duration / 16))
         #         last_size = int(math.ceil(sample_size / 32))
         # print(last_duration, last_size)
@@ -167,13 +141,13 @@ class ResNet(nn.Module):
         #         self.avgpool = nn.AvgPool3d(
         #             (last_duration, last_size, last_size), stride=1)
         self.dropout_p = dropout_p
-        if dropout_p>0:
+        if dropout_p > 0:
             self.dropout = torch.nn.Dropout(p=dropout_p)
         self.fc = nn.Linear(512 * block.expansion, num_classes)
-        
+
         for m in self.modules():
             if isinstance(m, nn.Conv3d):
-                m.weight = nn.init.kaiming_normal(m.weight, mode='fan_out')
+                m.weight = nn.init.kaiming_normal(m.weight, mode="fan_out")
             elif isinstance(m, nn.BatchNorm3d):
                 m.weight.data.fill_(1)
                 m.bias.data.zero_()
@@ -181,19 +155,13 @@ class ResNet(nn.Module):
     def _make_layer(self, block, planes, blocks, shortcut_type, stride=1):
         downsample = None
         if stride != 1 or self.inplanes != planes * block.expansion:
-            if shortcut_type == 'A':
-                downsample = partial(
-                    downsample_basic_block,
-                    planes=planes * block.expansion,
-                    stride=stride)
+            if shortcut_type == "A":
+                downsample = partial(downsample_basic_block, planes=planes * block.expansion, stride=stride)
             else:
                 downsample = nn.Sequential(
-                    nn.Conv3d(
-                        self.inplanes,
-                        planes * block.expansion,
-                        kernel_size=1,
-                        stride=stride,
-                        bias=False), nn.BatchNorm3d(planes * block.expansion))
+                    nn.Conv3d(self.inplanes, planes * block.expansion, kernel_size=1, stride=stride, bias=False),
+                    nn.BatchNorm3d(planes * block.expansion),
+                )
 
         layers = []
         layers.append(block(self.inplanes, planes, stride, downsample))
@@ -218,10 +186,10 @@ class ResNet(nn.Module):
         x = self.avgpool(x)
 
         x = x.view(x.size(0), -1)
-        if self.dropout_p>0:
+        if self.dropout_p > 0:
             x = self.dropout(x)
         x = self.fc(x)
-        
+
         return x
 
 
@@ -231,73 +199,66 @@ def get_fine_tuning_parameters(model, ft_begin_index):
 
     ft_module_names = []
     for i in range(ft_begin_index, 5):
-        ft_module_names.append('layer{}'.format(i))
-    ft_module_names.append('fc')
+        ft_module_names.append("layer{}".format(i))
+    ft_module_names.append("fc")
 
     parameters = []
     for k, v in model.named_parameters():
         for ft_module in ft_module_names:
             if ft_module in k:
-                parameters.append({'params': v})
+                parameters.append({"params": v})
                 break
         else:
-            parameters.append({'params': v, 'lr': 0.0})
+            parameters.append({"params": v, "lr": 0.0})
 
     return parameters
 
 
 def resnet3d_10(**kwargs):
-    """Constructs a ResNet-18 model.
-    """
+    """Constructs a ResNet-18 model."""
     model = ResNet(BasicBlock, [1, 1, 1, 1], **kwargs)
     return model
 
 
 def resnet3d_18(**kwargs):
-    """Constructs a ResNet-18 model.
-    """
+    """Constructs a ResNet-18 model."""
     model = ResNet(BasicBlock, [2, 2, 2, 2], **kwargs)
     return model
 
 
 def resnet3d_34(pretrained=False, in_channels=3, path_to_weights=None, **kwargs):
-    """Constructs a ResNet-34 model.
-    """
+    """Constructs a ResNet-34 model."""
     model = ResNet(BasicBlock, [3, 4, 6, 3], in_channels=in_channels, **kwargs)
     if pretrained:
         if in_channels != 3:
-            warnings.warn('in channels is {}, not reloading imagenet weights...'.format(in_channels))
+            warnings.warn("in channels is {}, not reloading imagenet weights...".format(in_channels))
         else:
-            warnings.warn('Using absolute file import for resnet3d weights')
+            warnings.warn("Using absolute file import for resnet3d weights")
             if path_to_weights is None:
-                raise ValueError('must specify path to weights file if pretrained: {}'.format(path_to_weights))
+                raise ValueError("must specify path to weights file if pretrained: {}".format(path_to_weights))
             model, _, _, _ = load_state(model, path_to_weights)
     return model
 
 
 def resnet3d_50(**kwargs):
-    """Constructs a ResNet-50 model.
-    """
+    """Constructs a ResNet-50 model."""
     model = ResNet(Bottleneck, [3, 4, 6, 3], **kwargs)
     return model
 
 
 def resnet3d_101(**kwargs):
-    """Constructs a ResNet-101 model.
-    """
+    """Constructs a ResNet-101 model."""
     model = ResNet(Bottleneck, [3, 4, 23, 3], **kwargs)
     return model
 
 
 def resnet3d_152(**kwargs):
-    """Constructs a ResNet-101 model.
-    """
+    """Constructs a ResNet-101 model."""
     model = ResNet(Bottleneck, [3, 8, 36, 3], **kwargs)
     return model
 
 
 def resnet3d_200(**kwargs):
-    """Constructs a ResNet-101 model.
-    """
+    """Constructs a ResNet-101 model."""
     model = ResNet(Bottleneck, [3, 24, 36, 3], **kwargs)
     return model
