@@ -36,45 +36,38 @@ import torch.nn.init as init
 import torch.utils.model_zoo as model_zoo
 
 
-__all__ = ['SqueezeNet', 'squeezenet1_0', 'squeezenet1_1']
+__all__ = ["SqueezeNet", "squeezenet1_0", "squeezenet1_1"]
 
 
 model_urls = {
-    'squeezenet1_0': 'https://download.pytorch.org/models/squeezenet1_0-a815701f.pth',
-    'squeezenet1_1': 'https://download.pytorch.org/models/squeezenet1_1-f364aa15.pth',
+    "squeezenet1_0": "https://download.pytorch.org/models/squeezenet1_0-a815701f.pth",
+    "squeezenet1_1": "https://download.pytorch.org/models/squeezenet1_1-f364aa15.pth",
 }
 
 
 class Fire(nn.Module):
-
-    def __init__(self, inplanes, squeeze_planes,
-                 expand1x1_planes, expand3x3_planes):
+    def __init__(self, inplanes, squeeze_planes, expand1x1_planes, expand3x3_planes):
         super(Fire, self).__init__()
         self.inplanes = inplanes
         self.squeeze = nn.Conv2d(inplanes, squeeze_planes, kernel_size=1)
         self.squeeze_activation = nn.ReLU(inplace=True)
-        self.expand1x1 = nn.Conv2d(squeeze_planes, expand1x1_planes,
-                                   kernel_size=1)
+        self.expand1x1 = nn.Conv2d(squeeze_planes, expand1x1_planes, kernel_size=1)
         self.expand1x1_activation = nn.ReLU(inplace=True)
-        self.expand3x3 = nn.Conv2d(squeeze_planes, expand3x3_planes,
-                                   kernel_size=3, padding=1)
+        self.expand3x3 = nn.Conv2d(squeeze_planes, expand3x3_planes, kernel_size=3, padding=1)
         self.expand3x3_activation = nn.ReLU(inplace=True)
 
     def forward(self, x):
         x = self.squeeze_activation(self.squeeze(x))
-        return torch.cat([
-            self.expand1x1_activation(self.expand1x1(x)),
-            self.expand3x3_activation(self.expand3x3(x))
-        ], 1)
+        return torch.cat(
+            [self.expand1x1_activation(self.expand1x1(x)), self.expand3x3_activation(self.expand3x3(x))], 1
+        )
 
 
 class SqueezeNet(nn.Module):
-
-    def __init__(self, version=1.0, in_channels=3,num_classes=1000):
+    def __init__(self, version=1.0, in_channels=3, num_classes=1000):
         super(SqueezeNet, self).__init__()
         if version not in [1.0, 1.1]:
-            raise ValueError("Unsupported SqueezeNet version {version}:"
-                             "1.0 or 1.1 expected".format(version=version))
+            raise ValueError("Unsupported SqueezeNet version {version}:1.0 or 1.1 expected".format(version=version))
         self.num_classes = num_classes
         if version == 1.0:
             self.features = nn.Sequential(
@@ -111,10 +104,7 @@ class SqueezeNet(nn.Module):
         # Final convolution is initialized differently form the rest
         final_conv = nn.Conv2d(512, self.num_classes, kernel_size=1)
         self.classifier = nn.Sequential(
-            nn.Dropout(p=0.5),
-            final_conv,
-            nn.ReLU(inplace=True),
-            nn.AdaptiveAvgPool2d((1, 1))
+            nn.Dropout(p=0.5), final_conv, nn.ReLU(inplace=True), nn.AdaptiveAvgPool2d((1, 1))
         )
 
         for m in self.modules():
@@ -132,7 +122,7 @@ class SqueezeNet(nn.Module):
         return x.view(x.size(0), self.num_classes)
 
 
-def squeezenet1_0(pretrained=False,in_channels=3, **kwargs):
+def squeezenet1_0(pretrained=False, in_channels=3, **kwargs):
     r"""SqueezeNet model architecture from the `"SqueezeNet: AlexNet-level
     accuracy with 50x fewer parameters and <0.5MB model size"
     <https://arxiv.org/abs/1602.07360>`_ paper.
@@ -143,18 +133,18 @@ def squeezenet1_0(pretrained=False,in_channels=3, **kwargs):
     model = SqueezeNet(in_channels=in_channels, version=1.0, **kwargs)
     if pretrained:
         # from Wang et al. 2015: Towards good practices for very deep two-stream convnets
-        state_dict = model_zoo.load_url(model_urls['squeezenet1_0'])
-        if in_channels !=3:
+        state_dict = model_zoo.load_url(model_urls["squeezenet1_0"])
+        if in_channels != 3:
             rgb_kernel_key = list(state_dict.keys())[0]
             rgb_kernel = state_dict[rgb_kernel_key]
-            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels,1,1)
+            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels, 1, 1)
             state_dict[rgb_kernel_key] = flow_kernel
             state_dict.update(state_dict)
         model.load_state_dict(state_dict)
     return model
 
 
-def squeezenet1_1(pretrained=False, in_channels=3,**kwargs):
+def squeezenet1_1(pretrained=False, in_channels=3, **kwargs):
     r"""SqueezeNet 1.1 model from the `official SqueezeNet repo
     <https://github.com/DeepScale/SqueezeNet/tree/master/SqueezeNet_v1.1>`_.
     SqueezeNet 1.1 has 2.4x less computation and slightly fewer parameters
@@ -163,14 +153,14 @@ def squeezenet1_1(pretrained=False, in_channels=3,**kwargs):
     Args:
         pretrained (bool): If True, returns a model pre-trained on ImageNet
     """
-    model = SqueezeNet(in_channels=in_channels,version=1.1, **kwargs)
+    model = SqueezeNet(in_channels=in_channels, version=1.1, **kwargs)
     if pretrained:
         # from Wang et al. 2015: Towards good practices for very deep two-stream convnets
-        state_dict = model_zoo.load_url(model_urls['squeeznet1_1'])
-        if in_channels !=3:
+        state_dict = model_zoo.load_url(model_urls["squeeznet1_1"])
+        if in_channels != 3:
             rgb_kernel_key = list(state_dict.keys())[0]
             rgb_kernel = state_dict[rgb_kernel_key]
-            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels,1,1)
+            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels, 1, 1)
             state_dict[rgb_kernel_key] = flow_kernel
             state_dict.update(state_dict)
         model.load_state_dict(state_dict)
