@@ -33,25 +33,31 @@ import torch.nn as nn
 import torch.utils.model_zoo as model_zoo
 
 __all__ = [
-    'VGG', 'vgg11', 'vgg11_bn', 'vgg13', 'vgg13_bn', 'vgg16', 'vgg16_bn',
-    'vgg19_bn', 'vgg19',
+    "VGG",
+    "vgg11",
+    "vgg11_bn",
+    "vgg13",
+    "vgg13_bn",
+    "vgg16",
+    "vgg16_bn",
+    "vgg19_bn",
+    "vgg19",
 ]
 
 
 model_urls = {
-    'vgg11': 'https://download.pytorch.org/models/vgg11-bbd30ac9.pth',
-    'vgg13': 'https://download.pytorch.org/models/vgg13-c768596a.pth',
-    'vgg16': 'https://download.pytorch.org/models/vgg16-397923af.pth',
-    'vgg19': 'https://download.pytorch.org/models/vgg19-dcbb9e9d.pth',
-    'vgg11_bn': 'https://download.pytorch.org/models/vgg11_bn-6002323d.pth',
-    'vgg13_bn': 'https://download.pytorch.org/models/vgg13_bn-abd245e5.pth',
-    'vgg16_bn': 'https://download.pytorch.org/models/vgg16_bn-6c64b313.pth',
-    'vgg19_bn': 'https://download.pytorch.org/models/vgg19_bn-c79401a0.pth',
+    "vgg11": "https://download.pytorch.org/models/vgg11-bbd30ac9.pth",
+    "vgg13": "https://download.pytorch.org/models/vgg13-c768596a.pth",
+    "vgg16": "https://download.pytorch.org/models/vgg16-397923af.pth",
+    "vgg19": "https://download.pytorch.org/models/vgg19-dcbb9e9d.pth",
+    "vgg11_bn": "https://download.pytorch.org/models/vgg11_bn-6002323d.pth",
+    "vgg13_bn": "https://download.pytorch.org/models/vgg13_bn-abd245e5.pth",
+    "vgg16_bn": "https://download.pytorch.org/models/vgg16_bn-6c64b313.pth",
+    "vgg19_bn": "https://download.pytorch.org/models/vgg19_bn-c79401a0.pth",
 }
 
 
 class VGG(nn.Module):
-
     def __init__(self, features, num_classes=1000, init_weights=True, dropout_p=0):
         super(VGG, self).__init__()
         self.features = features
@@ -76,7 +82,7 @@ class VGG(nn.Module):
     def _initialize_weights(self):
         for m in self.modules():
             if isinstance(m, nn.Conv2d):
-                nn.init.kaiming_normal_(m.weight, mode='fan_out', nonlinearity='relu')
+                nn.init.kaiming_normal_(m.weight, mode="fan_out", nonlinearity="relu")
                 if m.bias is not None:
                     nn.init.constant_(m.bias, 0)
             elif isinstance(m, nn.BatchNorm2d):
@@ -87,10 +93,10 @@ class VGG(nn.Module):
                 nn.init.constant_(m.bias, 0)
 
 
-def make_layers(cfg, in_channels=3,batch_norm=False):
+def make_layers(cfg, in_channels=3, batch_norm=False):
     layers = []
     for v in cfg:
-        if v == 'M':
+        if v == "M":
             layers += [nn.MaxPool2d(kernel_size=2, stride=2)]
         else:
             conv2d = nn.Conv2d(in_channels, v, kernel_size=3, padding=1)
@@ -103,183 +109,183 @@ def make_layers(cfg, in_channels=3,batch_norm=False):
 
 
 cfg = {
-    'A': [64, 'M', 128, 'M', 256, 256, 'M', 512, 512, 'M', 512, 512, 'M'],
-    'B': [64, 64, 'M', 128, 128, 'M', 256, 256, 'M', 512, 512, 'M', 512, 512, 'M'],
-    'D': [64, 64, 'M', 128, 128, 'M', 256, 256, 256, 'M', 512, 512, 512, 'M', 512, 512, 512, 'M'],
-    'E': [64, 64, 'M', 128, 128, 'M', 256, 256, 256, 256, 'M', 512, 512, 512, 512, 'M', 512, 512, 512, 512, 'M'],
+    "A": [64, "M", 128, "M", 256, 256, "M", 512, 512, "M", 512, 512, "M"],
+    "B": [64, 64, "M", 128, 128, "M", 256, 256, "M", 512, 512, "M", 512, 512, "M"],
+    "D": [64, 64, "M", 128, 128, "M", 256, 256, 256, "M", 512, 512, 512, "M", 512, 512, 512, "M"],
+    "E": [64, 64, "M", 128, 128, "M", 256, 256, 256, 256, "M", 512, 512, 512, 512, "M", 512, 512, 512, 512, "M"],
 }
 
 
-def vgg11(pretrained=False, in_channels=3,**kwargs):
+def vgg11(pretrained=False, in_channels=3, **kwargs):
     """VGG 11-layer model (configuration "A")
 
     Args:
         pretrained (bool): If True, returns a model pre-trained on ImageNet
     """
     if pretrained:
-        kwargs['init_weights'] = False
-    model = VGG(make_layers(cfg['A'], in_channels=in_channels), **kwargs)
+        kwargs["init_weights"] = False
+    model = VGG(make_layers(cfg["A"], in_channels=in_channels), **kwargs)
     if pretrained:
         # from Wang et al. 2015: Towards good practices for very deep two-stream convnets
-        state_dict = model_zoo.load_url(model_urls['vgg11'])
-        if in_channels !=3:
+        state_dict = model_zoo.load_url(model_urls["vgg11"])
+        if in_channels != 3:
             rgb_kernel_key = list(state_dict.keys())[0]
             rgb_kernel = state_dict[rgb_kernel_key]
-            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels,1,1)
+            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels, 1, 1)
             state_dict[rgb_kernel_key] = flow_kernel
             state_dict.update(state_dict)
         model.load_state_dict(state_dict)
     return model
 
 
-def vgg11_bn(pretrained=False, in_channels=3,**kwargs):
+def vgg11_bn(pretrained=False, in_channels=3, **kwargs):
     """VGG 11-layer model (configuration "A") with batch normalization
 
     Args:
         pretrained (bool): If True, returns a model pre-trained on ImageNet
     """
     if pretrained:
-        kwargs['init_weights'] = False
-    model = VGG(make_layers(cfg['A'], in_channels=in_channels, batch_norm=True), **kwargs)
+        kwargs["init_weights"] = False
+    model = VGG(make_layers(cfg["A"], in_channels=in_channels, batch_norm=True), **kwargs)
     if pretrained:
         # from Wang et al. 2015: Towards good practices for very deep two-stream convnets
-        state_dict = model_zoo.load_url(model_urls['vgg11_bn'])
-        if in_channels !=3:
+        state_dict = model_zoo.load_url(model_urls["vgg11_bn"])
+        if in_channels != 3:
             rgb_kernel_key = list(state_dict.keys())[0]
             rgb_kernel = state_dict[rgb_kernel_key]
-            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels,1,1)
+            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels, 1, 1)
             state_dict[rgb_kernel_key] = flow_kernel
             state_dict.update(state_dict)
         model.load_state_dict(state_dict)
     return model
 
 
-def vgg13(pretrained=False, in_channels=3,**kwargs):
+def vgg13(pretrained=False, in_channels=3, **kwargs):
     """VGG 13-layer model (configuration "B")
 
     Args:
         pretrained (bool): If True, returns a model pre-trained on ImageNet
     """
     if pretrained:
-        kwargs['init_weights'] = False
-    model = VGG(make_layers(cfg['B'], in_channels=in_channels), **kwargs)
+        kwargs["init_weights"] = False
+    model = VGG(make_layers(cfg["B"], in_channels=in_channels), **kwargs)
     if pretrained:
         # from Wang et al. 2015: Towards good practices for very deep two-stream convnets
-        state_dict = model_zoo.load_url(model_urls['vgg13'])
-        if in_channels !=3:
+        state_dict = model_zoo.load_url(model_urls["vgg13"])
+        if in_channels != 3:
             rgb_kernel_key = list(state_dict.keys())[0]
             rgb_kernel = state_dict[rgb_kernel_key]
-            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels,1,1)
+            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels, 1, 1)
             state_dict[rgb_kernel_key] = flow_kernel
             state_dict.update(state_dict)
         model.load_state_dict(state_dict)
     return model
 
 
-def vgg13_bn(pretrained=False,in_channels=3,**kwargs):
+def vgg13_bn(pretrained=False, in_channels=3, **kwargs):
     """VGG 13-layer model (configuration "B") with batch normalization
 
     Args:
         pretrained (bool): If True, returns a model pre-trained on ImageNet
     """
     if pretrained:
-        kwargs['init_weights'] = False
-    model = VGG(make_layers(cfg['B'], in_channels=in_channels,batch_norm=True), **kwargs)
+        kwargs["init_weights"] = False
+    model = VGG(make_layers(cfg["B"], in_channels=in_channels, batch_norm=True), **kwargs)
     if pretrained:
         # from Wang et al. 2015: Towards good practices for very deep two-stream convnets
-        state_dict = model_zoo.load_url(model_urls['vgg13_bn'])
-        if in_channels !=3:
+        state_dict = model_zoo.load_url(model_urls["vgg13_bn"])
+        if in_channels != 3:
             rgb_kernel_key = list(state_dict.keys())[0]
             rgb_kernel = state_dict[rgb_kernel_key]
-            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels,1,1)
+            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels, 1, 1)
             state_dict[rgb_kernel_key] = flow_kernel
             state_dict.update(state_dict)
         model.load_state_dict(state_dict)
     return model
 
 
-def vgg16(pretrained=False, in_channels=3,**kwargs):
+def vgg16(pretrained=False, in_channels=3, **kwargs):
     """VGG 16-layer model (configuration "D")
 
     Args:
         pretrained (bool): If True, returns a model pre-trained on ImageNet
     """
     if pretrained:
-        kwargs['init_weights'] = False
-    model = VGG(make_layers(cfg['D'],in_channels=in_channels), **kwargs)
+        kwargs["init_weights"] = False
+    model = VGG(make_layers(cfg["D"], in_channels=in_channels), **kwargs)
     if pretrained:
         # from Wang et al. 2015: Towards good practices for very deep two-stream convnets
-        state_dict = model_zoo.load_url(model_urls['vgg16'])
-        if in_channels !=3:
+        state_dict = model_zoo.load_url(model_urls["vgg16"])
+        if in_channels != 3:
             rgb_kernel_key = list(state_dict.keys())[0]
             rgb_kernel = state_dict[rgb_kernel_key]
-            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels,1,1)
+            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels, 1, 1)
             state_dict[rgb_kernel_key] = flow_kernel
             state_dict.update(state_dict)
         model.load_state_dict(state_dict)
     return model
 
 
-def vgg16_bn(pretrained=False,in_channels=3,**kwargs):
+def vgg16_bn(pretrained=False, in_channels=3, **kwargs):
     """VGG 16-layer model (configuration "D") with batch normalization
 
     Args:
         pretrained (bool): If True, returns a model pre-trained on ImageNet
     """
     if pretrained:
-        kwargs['init_weights'] = False
-    model = VGG(make_layers(cfg['D'],in_channels=in_channels,batch_norm=True), **kwargs)
+        kwargs["init_weights"] = False
+    model = VGG(make_layers(cfg["D"], in_channels=in_channels, batch_norm=True), **kwargs)
     if pretrained:
         # from Wang et al. 2015: Towards good practices for very deep two-stream convnets
-        state_dict = model_zoo.load_url(model_urls['vgg16_bn'])
-        if in_channels !=3:
+        state_dict = model_zoo.load_url(model_urls["vgg16_bn"])
+        if in_channels != 3:
             rgb_kernel_key = list(state_dict.keys())[0]
             rgb_kernel = state_dict[rgb_kernel_key]
-            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels,1,1)
+            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels, 1, 1)
             state_dict[rgb_kernel_key] = flow_kernel
             state_dict.update(state_dict)
         model.load_state_dict(state_dict)
     return model
 
 
-def vgg19(pretrained=False,in_channels=3, **kwargs):
+def vgg19(pretrained=False, in_channels=3, **kwargs):
     """VGG 19-layer model (configuration "E")
 
     Args:
         pretrained (bool): If True, returns a model pre-trained on ImageNet
     """
     if pretrained:
-        kwargs['init_weights'] = False
-    model = VGG(make_layers(cfg['E'],in_channels=in_channels), **kwargs)
+        kwargs["init_weights"] = False
+    model = VGG(make_layers(cfg["E"], in_channels=in_channels), **kwargs)
     if pretrained:
         # from Wang et al. 2015: Towards good practices for very deep two-stream convnets
-        state_dict = model_zoo.load_url(model_urls['vgg19'])
-        if in_channels !=3:
+        state_dict = model_zoo.load_url(model_urls["vgg19"])
+        if in_channels != 3:
             rgb_kernel_key = list(state_dict.keys())[0]
             rgb_kernel = state_dict[rgb_kernel_key]
-            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels,1,1)
+            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels, 1, 1)
             state_dict[rgb_kernel_key] = flow_kernel
             state_dict.update(state_dict)
         model.load_state_dict(state_dict)
     return model
 
 
-def vgg19_bn(pretrained=False, in_channels=3,**kwargs):
+def vgg19_bn(pretrained=False, in_channels=3, **kwargs):
     """VGG 19-layer model (configuration 'E') with batch normalization
 
     Args:
         pretrained (bool): If True, returns a model pre-trained on ImageNet
     """
     if pretrained:
-        kwargs['init_weights'] = False
-    model = VGG(make_layers(cfg['E'], in_channels=in_channels,batch_norm=True), **kwargs)
+        kwargs["init_weights"] = False
+    model = VGG(make_layers(cfg["E"], in_channels=in_channels, batch_norm=True), **kwargs)
     if pretrained:
         # from Wang et al. 2015: Towards good practices for very deep two-stream convnets
-        state_dict = model_zoo.load_url(model_urls['vgg19_bn'])
-        if in_channels !=3:
+        state_dict = model_zoo.load_url(model_urls["vgg19_bn"])
+        if in_channels != 3:
             rgb_kernel_key = list(state_dict.keys())[0]
             rgb_kernel = state_dict[rgb_kernel_key]
-            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels,1,1)
+            flow_kernel = rgb_kernel.mean(dim=1).unsqueeze(1).repeat(1, in_channels, 1, 1)
             state_dict[rgb_kernel_key] = flow_kernel
             state_dict.update(state_dict)
         model.load_state_dict(state_dict)
