@@ -15,8 +15,9 @@ flow_generators = utils.get_models_from_module(flow_models, get_function=False)
 
 log = logging.getLogger(__name__)
 
+
 class Viewer(nn.Module):
-    """ PyTorch module for extracting the middle image of a concatenated stack.
+    """PyTorch module for extracting the middle image of a concatenated stack.
 
     Example: you have 10 RGB images stacked in a channel of a tensor, so it has shape [N, 30, H, W].
         viewer = Viewer(10)
@@ -28,19 +29,19 @@ class Viewer(nn.Module):
     def __init__(self, num_images, label_location):
         super().__init__()
         self.num_images = num_images
-        if label_location == 'middle':
+        if label_location == "middle":
             self.start = int(num_images / 2 * 3)
-        elif label_location == 'causal':
+        elif label_location == "causal":
             self.start = int(num_images * 3 - 3)
         self.end = int(self.start + 3)
 
     def forward(self, x):
-        x = x[:, self.start:self.end, :, :]
+        x = x[:, self.start : self.end, :, :]
         return x
 
 
 class FlowOnlyClassifier(nn.Module):
-    """ Stack of flow generator module and flow classifier. Used in training Hidden two stream networks in a curriculum
+    """Stack of flow generator module and flow classifier. Used in training Hidden two stream networks in a curriculum
 
     Takes a stack of images as inputs. Generates optic flow using the flow generator; computes class probabilities
     using the flow classifier.
@@ -55,11 +56,9 @@ class FlowOnlyClassifier(nn.Module):
         print(outputs.shape) # [N, K]
     """
 
-    def __init__(self, flow_generator,
-                 flow_classifier,
-                 freeze_flow_generator: bool = True):
+    def __init__(self, flow_generator, flow_classifier, freeze_flow_generator: bool = True):
         super().__init__()
-        assert (isinstance(flow_generator, nn.Module) and isinstance(flow_classifier, nn.Module))
+        assert isinstance(flow_generator, nn.Module) and isinstance(flow_classifier, nn.Module)
 
         self.flow_generator = flow_generator
         if freeze_flow_generator:
@@ -78,7 +77,7 @@ class FlowOnlyClassifier(nn.Module):
 
 
 class HiddenTwoStream(nn.Module):
-    """ Hidden Two-Stream Network model
+    """Hidden Two-Stream Network model
     Paper: https://arxiv.org/abs/1704.00389
 
     Classifies video inputs using a spatial CNN, using RGB video frames as inputs; and using a flow CNN, which
@@ -86,10 +85,17 @@ class HiddenTwoStream(nn.Module):
     advantages, as optic flow loaded from disk is both more discrete and has compression artifacts.
     """
 
-    def __init__(self, flow_generator, spatial_classifier, flow_classifier, fusion,
-                 classifier_name: str, num_images: int = 11,
-                 label_location: str = 'middle'):
-        """ Hidden two-stream constructor.
+    def __init__(
+        self,
+        flow_generator,
+        spatial_classifier,
+        flow_classifier,
+        fusion,
+        classifier_name: str,
+        num_images: int = 11,
+        label_location: str = "middle",
+    ):
+        """Hidden two-stream constructor.
 
         Args:
             flow_generator (nn.Module): CNN that generates optic flow from a stack of RGB frames
@@ -102,89 +108,93 @@ class HiddenTwoStream(nn.Module):
 
         """
         super().__init__()
-        assert (isinstance(flow_generator, nn.Module) and isinstance(spatial_classifier, nn.Module)
-                and isinstance(flow_classifier, nn.Module) and isinstance(fusion, nn.Module))
+        assert (
+            isinstance(flow_generator, nn.Module)
+            and isinstance(spatial_classifier, nn.Module)
+            and isinstance(flow_classifier, nn.Module)
+            and isinstance(fusion, nn.Module)
+        )
 
         self.spatial_classifier = spatial_classifier
         self.flow_generator = flow_generator
         self.flow_classifier = flow_classifier
-        if '3d' in classifier_name:
+        if "3d" in classifier_name:
             self.viewer = nn.Identity()
         else:
             self.viewer = Viewer(num_images, label_location)
         self.fusion = fusion
 
         self.frozen_state = {}
-        self.freeze('flow_generator')
+        self.freeze("flow_generator")
 
     def freeze(self, submodel_to_freeze: str):
-        """ Freezes a component of the model. Useful for curriculum training
+        """Freezes a component of the model. Useful for curriculum training
 
         Args:
             submodel_to_freeze (str): one of flow_generator, spatial, flow, fusion
         """
-        if submodel_to_freeze == 'flow_generator':
+        if submodel_to_freeze == "flow_generator":
             self.flow_generator.eval()
             for param in self.flow_generator.parameters():
                 param.requires_grad = False
-        elif submodel_to_freeze == 'spatial':
+        elif submodel_to_freeze == "spatial":
             self.spatial_classifier.eval()
             for param in self.spatial_classifier.parameters():
                 param.requires_grad = False
-        elif submodel_to_freeze == 'flow':
+        elif submodel_to_freeze == "flow":
             self.flow_classifier.eval()
             for param in self.flow_classifier.parameters():
                 param.requires_grad = False
-        elif submodel_to_freeze == 'fusion':
+        elif submodel_to_freeze == "fusion":
             self.fusion.eval()
             for param in self.fusion.parameters():
                 param.requires_grad = False
         else:
-            raise ValueError('submodel not found:%s' % submodel_to_freeze)
+            raise ValueError("submodel not found:%s" % submodel_to_freeze)
         self.frozen_state[submodel_to_freeze] = True
 
     def set_mode(self, mode: str):
-        """ Freezes and unfreezes portions of the model, useful for curriculum training.
+        """Freezes and unfreezes portions of the model, useful for curriculum training.
 
         Args:
             mode (str): one of spatial_only, flow_only, fusion_only, classifier, end_to_end, or inference
         """
-        log.debug('setting model mode: {}'.format(mode))
-        if mode == 'spatial_only':
-            self.freeze('flow_generator')
-            self.freeze('flow')
-            self.freeze('fusion')
-            self.unfreeze('spatial')
-        elif mode == 'flow_only':
-            self.freeze('flow_generator')
-            self.freeze('spatial')
-            self.unfreeze('flow')
-            self.freeze('fusion')
-        elif mode == 'fusion_only':
-            self.freeze('flow_generator')
-            self.freeze('spatial')
-            self.freeze('flow')
-            self.unfreeze('fusion')
-        elif mode == 'classifier':
-            self.freeze('flow_generator')
-            self.unfreeze('spatial')
-            self.unfreeze('flow')
-            self.unfreeze('fusion')
-        elif mode == 'end_to_end':
-            self.unfreeze('flow_generator')
-            self.unfreeze('spatial')
-            self.unfreeze('flow')
-            self.unfreeze('fusion')
-        elif mode == 'inference':
-            self.freeze('flow_generator')
-            self.freeze('spatial')
-            self.freeze('flow')
-            self.freeze('fusion')
+        log.debug("setting model mode: {}".format(mode))
+        if mode == "spatial_only":
+            self.freeze("flow_generator")
+            self.freeze("flow")
+            self.freeze("fusion")
+            self.unfreeze("spatial")
+        elif mode == "flow_only":
+            self.freeze("flow_generator")
+            self.freeze("spatial")
+            self.unfreeze("flow")
+            self.freeze("fusion")
+        elif mode == "fusion_only":
+            self.freeze("flow_generator")
+            self.freeze("spatial")
+            self.freeze("flow")
+            self.unfreeze("fusion")
+        elif mode == "classifier":
+            self.freeze("flow_generator")
+            self.unfreeze("spatial")
+            self.unfreeze("flow")
+            self.unfreeze("fusion")
+        elif mode == "end_to_end":
+            self.unfreeze("flow_generator")
+            self.unfreeze("spatial")
+            self.unfreeze("flow")
+            self.unfreeze("fusion")
+        elif mode == "inference":
+            self.freeze("flow_generator")
+            self.freeze("spatial")
+            self.freeze("flow")
+            self.freeze("fusion")
         else:
-            raise ValueError('Unknown mode: %s' % mode)
+            raise ValueError("Unknown mode: %s" % mode)
 
     def unfreeze(self, submodel_to_unfreeze: str):
-        """ Unfreezes portions of the model
+        """Unfreezes portions of the model
 
         Args:
             submodel_to_unfreeze (str): one of flow_generator, spatial, flow, or fusion
@@ -192,33 +202,35 @@ class HiddenTwoStream(nn.Module):
         Returns:
 
         """
-        log.debug('unfreezing model component: {}'.format(submodel_to_unfreeze))
-        if submodel_to_unfreeze == 'flow_generator':
+        log.debug("unfreezing model component: {}".format(submodel_to_unfreeze))
+        if submodel_to_unfreeze == "flow_generator":
             self.flow_generator.train()
             for param in self.flow_generator.parameters():
                 param.requires_grad = True
-        elif submodel_to_unfreeze == 'spatial':
+        elif submodel_to_unfreeze == "spatial":
             self.spatial_classifier.train()
             for param in self.spatial_classifier.parameters():
                 param.requires_grad = True
-        elif submodel_to_unfreeze == 'flow':
+        elif submodel_to_unfreeze == "flow":
             self.flow_classifier.train()
             for param in self.flow_classifier.parameters():
                 param.requires_grad = True
-        elif submodel_to_unfreeze == 'fusion':
+        elif submodel_to_unfreeze == "fusion":
             self.fusion.train()
             for param in self.fusion.parameters():
                 param.requires_grad = True
         else:
-            raise ValueError('submodel not found:%s' % submodel_to_unfreeze)
+            raise ValueError("submodel not found:%s" % submodel_to_unfreeze)
         self.frozen_state[submodel_to_unfreeze] = False
 
     def get_param_groups(self):
-        param_list = [{'params': self.flow_generator.parameters()},
-                      {'params': self.spatial_classifier.parameters()},
-                      {'params': self.flow_classifier.parameters()},
-                      {'params': self.fusion.parameters()}]
-        return (param_list)
+        param_list = [
+            {"params": self.flow_generator.parameters()},
+            {"params": self.spatial_classifier.parameters()},
+            {"params": self.flow_classifier.parameters()},
+            {"params": self.fusion.parameters()},
+        ]
+        return param_list
 
     def forward(self, batch):
         with torch.no_grad():
@@ -230,19 +242,21 @@ class HiddenTwoStream(nn.Module):
         return self.fusion(spatial_features, flow_features)
 
 
-def hidden_two_stream(classifier: str,
-                      flow_gen: str,
-                      num_classes: int,
-                      fusion_style: str = 'average',
-                      dropout_p: float = 0.9,
-                      reload_imagenet: bool = True,
-                      num_rgb: int = 1,
-                      num_flows: int = 10,
-                      pos: np.ndarray = None,
-                      neg: np.ndarray = None,
-                      flow_max: float = 5.0,
-                      **kwargs):
-    """ Wrapper for initializing hidden two stream models
+def hidden_two_stream(
+    classifier: str,
+    flow_gen: str,
+    num_classes: int,
+    fusion_style: str = "average",
+    dropout_p: float = 0.9,
+    reload_imagenet: bool = True,
+    num_rgb: int = 1,
+    num_flows: int = 10,
+    pos: np.ndarray = None,
+    neg: np.ndarray = None,
+    flow_max: float = 5.0,
+    **kwargs,
+):
+    """Wrapper for initializing hidden two stream models
 
     Args:
         classifier (str): a supported classifier, e.g. resnet18, vgg16
@@ -260,22 +274,37 @@ def hidden_two_stream(classifier: str,
     Returns:
         hidden two stream network model
     """
-    assert fusion_style in ['average', 'concatenate']
+    assert fusion_style in ["average", "concatenate"]
 
     flow_generator = flow_generators[flow_gen](num_images=num_flows + 1, flow_div=flow_max)
 
-    in_channels = num_rgb * 3 if '3d' not in classifier.lower() else 3
-    spatial_classifier = get_cnn(classifier, in_channels=in_channels, dropout_p=dropout_p,
-                                 num_classes=num_classes, reload_imagenet=reload_imagenet,
-                                 pos=pos, neg=neg, **kwargs)
+    in_channels = num_rgb * 3 if "3d" not in classifier.lower() else 3
+    spatial_classifier = get_cnn(
+        classifier,
+        in_channels=in_channels,
+        dropout_p=dropout_p,
+        num_classes=num_classes,
+        reload_imagenet=reload_imagenet,
+        pos=pos,
+        neg=neg,
+        **kwargs,
+    )
 
-    in_channels = num_flows * 2 if '3d' not in classifier.lower() else 2
-    flow_classifier = get_cnn(classifier, in_channels=in_channels, dropout_p=dropout_p,
-                              num_classes=num_classes, reload_imagenet=reload_imagenet,
-                              pos=pos, neg=neg, **kwargs)
+    in_channels = num_flows * 2 if "3d" not in classifier.lower() else 2
+    flow_classifier = get_cnn(
+        classifier,
+        in_channels=in_channels,
+        dropout_p=dropout_p,
+        num_classes=num_classes,
+        reload_imagenet=reload_imagenet,
+        pos=pos,
+        neg=neg,
+        **kwargs,
+    )
 
-    spatial_classifier, flow_classifier, fusion = build_fusion_layer(spatial_classifier, flow_classifier,
-                                                                     fusion_style, num_classes)
+    spatial_classifier, flow_classifier, fusion = build_fusion_layer(
+        spatial_classifier, flow_classifier, fusion_style, num_classes
+    )
 
     model = HiddenTwoStream(flow_generator, spatial_classifier, flow_classifier, fusion, classifier)
     return model
@@ -296,23 +325,23 @@ def build_fusion_layer(spatial_classifier, flow_classifier, fusion_style, num_cl
     Returns:
 
     """
-    if fusion_style == 'average' or fusion_style == 'weighted_average':
+    if fusion_style == "average" or fusion_style == "weighted_average":
         # just so we can pass them to the fusion module
         num_spatial_features, num_flow_features = None, None
-    elif fusion_style == 'concatenate':
+    elif fusion_style == "concatenate":
         spatial_classifier, num_spatial_features = remove_cnn_classifier_layer(spatial_classifier)
         flow_classifier, num_flow_features = remove_cnn_classifier_layer(flow_classifier)
     else:
-        raise ValueError('unknown fusion style: {}'.format(fusion_style))
+        raise ValueError("unknown fusion style: {}".format(fusion_style))
 
-    fusion = Fusion(fusion_style, num_spatial_features, num_flow_features, num_classes,
-                    flow_fusion_weight=flow_fusion_weight)
+    fusion = Fusion(
+        fusion_style, num_spatial_features, num_flow_features, num_classes, flow_fusion_weight=flow_fusion_weight
+    )
     return spatial_classifier, flow_classifier, fusion
 
 
-def deg_f(num_classes: int, dropout_p: float = 0.9, reload_imagenet: bool = True,
-          pos: int = None, neg: int = None):
-    """ Make the DEG-fast model. Uses ResNet18 for classification, TinyMotionNet for flow generation.
+def deg_f(num_classes: int, dropout_p: float = 0.9, reload_imagenet: bool = True, pos: int = None, neg: int = None):
+    """Make the DEG-fast model. Uses ResNet18 for classification, TinyMotionNet for flow generation.
     Number of flows: 10
     Number of RGB frames for classification: 1
 
@@ -325,26 +354,29 @@ def deg_f(num_classes: int, dropout_p: float = 0.9, reload_imagenet: bool = True
     Returns:
         DEG-f model
     """
-    classifier = 'resnet18'
-    flow_gen = 'TinyMotionNet'
+    classifier = "resnet18"
+    flow_gen = "TinyMotionNet"
     num_flows = 10
     num_rgb = 1
-    fusion_style = 'average'
-    model = hidden_two_stream(classifier, flow_gen, num_classes,
-                              fusion_style=fusion_style,
-                              dropout_p=dropout_p,
-                              reload_imagenet=reload_imagenet,
-                              num_rgb=num_rgb,
-                              num_flows=num_flows,
-                              pos=pos,
-                              neg=neg)
+    fusion_style = "average"
+    model = hidden_two_stream(
+        classifier,
+        flow_gen,
+        num_classes,
+        fusion_style=fusion_style,
+        dropout_p=dropout_p,
+        reload_imagenet=reload_imagenet,
+        num_rgb=num_rgb,
+        num_flows=num_flows,
+        pos=pos,
+        neg=neg,
+    )
 
     return model
 
 
-def deg_m(num_classes: int, dropout_p: float = 0.9, reload_imagenet: bool = True,
-          pos: int = None, neg: int = None):
-    """ Make the DEG-medium model. Uses ResNet50 for classification, MotionNet for flow generation.
+def deg_m(num_classes: int, dropout_p: float = 0.9, reload_imagenet: bool = True, pos: int = None, neg: int = None):
+    """Make the DEG-medium model. Uses ResNet50 for classification, MotionNet for flow generation.
     Number of flows: 10
     Number of RGB frames for classification: 1
 
@@ -357,26 +389,36 @@ def deg_m(num_classes: int, dropout_p: float = 0.9, reload_imagenet: bool = True
     Returns:
         DEG-m model
     """
-    classifier = 'resnet50'
-    flow_gen = 'MotionNet'
+    classifier = "resnet50"
+    flow_gen = "MotionNet"
     num_flows = 10
     num_rgb = 1
-    fusion_style = 'average'
-    model = hidden_two_stream(classifier, flow_gen, num_classes,
-                              fusion_style=fusion_style,
-                              dropout_p=dropout_p,
-                              reload_imagenet=reload_imagenet,
-                              num_rgb=num_rgb,
-                              num_flows=num_flows,
-                              pos=pos,
-                              neg=neg)
+    fusion_style = "average"
+    model = hidden_two_stream(
+        classifier,
+        flow_gen,
+        num_classes,
+        fusion_style=fusion_style,
+        dropout_p=dropout_p,
+        reload_imagenet=reload_imagenet,
+        num_rgb=num_rgb,
+        num_flows=num_flows,
+        pos=pos,
+        neg=neg,
+    )
 
     return model
 
 
-def deg_s(num_classes: int, dropout_p: float = 0.9, reload_imagenet: bool = True,
-          pos: int = None, neg: int = None, path_to_weights: Union[str, os.PathLike] = None):
-    """ Make the DEG-slow model. Uses ResNet3d-34 for classification, TinyMotionNet3D for flow generation.
+def deg_s(
+    num_classes: int,
+    dropout_p: float = 0.9,
+    reload_imagenet: bool = True,
+    pos: int = None,
+    neg: int = None,
+    path_to_weights: Union[str, os.PathLike] = None,
+):
+    """Make the DEG-slow model. Uses ResNet3d-34 for classification, TinyMotionNet3D for flow generation.
     Number of flows: 10
     Number of RGB frames for classification: 11
 
@@ -391,19 +433,23 @@ def deg_s(num_classes: int, dropout_p: float = 0.9, reload_imagenet: bool = True
     Returns:
         DEG-s model
     """
-    classifier = 'resnet3d_34'
-    flow_gen = 'TinyMotionNet3D'
+    classifier = "resnet3d_34"
+    flow_gen = "TinyMotionNet3D"
     num_flows = 10
     num_rgb = 11
-    fusion_style = 'average'
-    model = hidden_two_stream(classifier, flow_gen, num_classes,
-                              fusion_style=fusion_style,
-                              dropout_p=dropout_p,
-                              reload_imagenet=reload_imagenet,
-                              num_rgb=num_rgb,
-                              num_flows=num_flows,
-                              pos=pos,
-                              neg=neg,
-                              path_to_weights=path_to_weights)
+    fusion_style = "average"
+    model = hidden_two_stream(
+        classifier,
+        flow_gen,
+        num_classes,
+        fusion_style=fusion_style,
+        dropout_p=dropout_p,
+        reload_imagenet=reload_imagenet,
+        num_rgb=num_rgb,
+        num_flows=num_flows,
+        pos=pos,
+        neg=neg,
+        path_to_weights=path_to_weights,
+    )
 
     return model
