@@ -28,10 +28,13 @@ References
 .. [1]: Fischer et al. FlowNet: Learning optical flow with convolutional networks. ICCV 2015
         https://arxiv.org/abs/1504.06852
 """
+
+import torch
+import torch.nn as nn
 import torch.nn.functional as F
 from torch.nn import init
 
-from .components import *
+from .components import conv, deconv, get_hw, predict_flow
 
 
 class FlowNetS(nn.Module):
@@ -67,15 +70,7 @@ class FlowNetS(nn.Module):
         self.upsampled_flow4_to_3 = nn.ConvTranspose2d(2, 2, 4, 2, 1, bias=False)
         self.upsampled_flow3_to_2 = nn.ConvTranspose2d(2, 2, 4, 2, 1, bias=False)
 
-        # self.upsampled_flow6_to_5 = nn.Sequential(Interpolate(scale_factor=2),
-        #     nn.Conv2d(2,2,kernel_size=3, stride=1, padding=1, bias=False))
-        # self.upsampled_flow5_to_4 = nn.Sequential(Interpolate(scale_factor=2),
-        #     nn.Conv2d(2,2,kernel_size=3, stride=1, padding=1, bias=False))
-        # self.upsampled_flow4_to_3 = nn.Sequential(Interpolate(scale_factor=2),
-        #     nn.Conv2d(2,2,kernel_size=3, stride=1, padding=1, bias=False))
-        # self.upsampled_flow3_to_2 = nn.Sequential(Interpolate(scale_factor=2),
-        #     nn.Conv2d(2,2,kernel_size=3, stride=1, padding=1, bias=False))
-
+        # initialize weights
         for m in self.modules():
             if isinstance(m, nn.Conv2d):
                 if m.bias is not None:
@@ -86,8 +81,8 @@ class FlowNetS(nn.Module):
                 if m.bias is not None:
                     init.uniform_(m.bias)
                 init.xavier_uniform_(m.weight)
-                # init_deconv_bilinear(m.weight)
-        self.upsample1 = nn.Upsample(scale_factor=4, mode='bilinear')
+
+        self.upsample1 = nn.Upsample(scale_factor=4, mode="bilinear")
 
     def forward(self, x):
         out_conv1 = self.conv1(x)
@@ -112,26 +107,25 @@ class FlowNetS(nn.Module):
         # a value of 1 in flow6 will naively be mapped to a value of 1 in flow5. now, this movement of 1 pixel no
         # longer means 1/8 of the image, it will only move 1/16 of the image. So to correct for this, we multiply
         # the upsampled version by 2.
-        flow6_up = self.upsampled_flow6_to_5(flow6)*2
+        flow6_up = self.upsampled_flow6_to_5(flow6) * 2
         out_deconv5 = self.deconv5(out_conv6)
 
         concat5 = torch.cat((out_conv5, out_deconv5, flow6_up), 1)
         flow5 = self.predict_flow5(concat5) * self.flow_div
-        flow5_up = self.upsampled_flow5_to_4(flow5)*2
+        flow5_up = self.upsampled_flow5_to_4(flow5) * 2
         out_deconv4 = self.deconv4(concat5)
 
         concat4 = torch.cat((out_conv4, out_deconv4, flow5_up), 1)
         flow4 = self.predict_flow4(concat4) * self.flow_div
-        flow4_up = self.upsampled_flow4_to_3(flow4)*2
+        flow4_up = self.upsampled_flow4_to_3(flow4) * 2
         out_deconv3 = self.deconv3(concat4)
 
         if get_hw(out_conv3) != get_hw(out_deconv3):
-            out_conv3 = F.interpolate(out_conv3, size=get_hw(out_deconv3),
-                                      mode='bilinear', align_corners=False)
+            out_conv3 = F.interpolate(out_conv3, size=get_hw(out_deconv3), mode="bilinear", align_corners=False)
 
         concat3 = torch.cat((out_conv3, out_deconv3, flow4_up), 1)
         flow3 = self.predict_flow3(concat3) * self.flow_div
-        flow3_up = self.upsampled_flow3_to_2(flow3)*2
+        flow3_up = self.upsampled_flow3_to_2(flow3) * 2
         out_deconv2 = self.deconv2(concat3)
 
         concat2 = torch.cat((out_conv2, out_deconv2, flow3_up), 1)
@@ -140,4 +134,4 @@ class FlowNetS(nn.Module):
         if self.training:
             return flow2, flow3, flow4, flow5, flow6
         else:
-            return flow2,
+            return (flow2,)
