@@ -1,26 +1,25 @@
-from collections import defaultdict
 import logging
 import os
-from typing import Dict, Type, Tuple
+from collections import defaultdict
+from typing import Tuple, Type
 
 import h5py
 import numpy as np
-from omegaconf import DictConfig, OmegaConf
 import pandas as pd
+from omegaconf import DictConfig, OmegaConf
 
-from deepethogram import projects, file_io
+from deepethogram import file_io, projects
 
 log = logging.getLogger(__name__)
 
 
-def remove_low_thresholds(thresholds: np.ndarray,
-                          minimum: float = 0.01,
-                          f1s: np.ndarray = None,
-                          minimum_f1: float = 0.05) -> np.ndarray:
-    """ Replaces thresholds below a certain value with 0.5
-    
-    If the model completely fails, the optimum threshold might be something erreoneous, such as 
-    0.00001. This makes all predictions==1. 
+def remove_low_thresholds(
+    thresholds: np.ndarray, minimum: float = 0.01, f1s: np.ndarray = None, minimum_f1: float = 0.05
+) -> np.ndarray:
+    """Replaces thresholds below a certain value with 0.5
+
+    If the model completely fails, the optimum threshold might be something erreoneous, such as
+    0.00001. This makes all predictions==1.
 
     Parameters
     ----------
@@ -37,18 +36,18 @@ def remove_low_thresholds(thresholds: np.ndarray,
     """
     if np.sum(thresholds < minimum) > 0:
         indices = np.where(thresholds < minimum)[0]
-        log.debug('thresholds {} too low, setting to {}'.format(thresholds[indices], minimum))
+        log.debug("thresholds {} too low, setting to {}".format(thresholds[indices], minimum))
         thresholds[thresholds < minimum] = minimum
     if f1s is not None:
         if np.sum(f1s < minimum_f1) > 0:
             indices = np.where(f1s < minimum_f1)[0]
-            log.debug('f1 {} too low, setting to 0.5'.format(f1s))
+            log.debug("f1 {} too low, setting to 0.5".format(f1s))
             thresholds[f1s < minimum_f1] = 0.5
     return thresholds
 
 
 def get_onsets_offsets(binary: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
-    """ Gets the onset and offset indices of a binary array.
+    """Gets the onset and offset indices of a binary array.
 
     Onset: index at which the array goes from 0 -> 1 (the index with the 1, not the 0)
     offset: index at which the array goes from 1 -> 0 (the index with the 0, not the 1)
@@ -78,26 +77,25 @@ def get_onsets_offsets(binary: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
 
 
 def get_bouts(ethogram: np.ndarray) -> list:
-    """ Get bouts from an ethogram. Uses 1->0 and 0->1 changes to define bout starts and stops """
+    """Get bouts from an ethogram. Uses 1->0 and 0->1 changes to define bout starts and stops"""
     K = ethogram.shape[1]
     stats = []
     for i in range(K):
         onsets, offsets = get_onsets_offsets(ethogram[:, i])
         stat = {
-            'N': len(onsets),
-            'lengths': np.array([offset - onset for (onset, offset) in zip(onsets, offsets)]),
-            'starts': onsets,
-            'ends': offsets
+            "N": len(onsets),
+            "lengths": np.array([offset - onset for (onset, offset) in zip(onsets, offsets)]),
+            "starts": onsets,
+            "ends": offsets,
         }
         stats.append(stat)
     return stats
 
 
-def find_bout_indices(predictions_trace: np.ndarray,
-                      bout_length: int,
-                      positive: bool = True,
-                      eps: float = 1e-6) -> np.ndarray:
-    """ Find indices where a bout of bout-length occurs in a binary vector
+def find_bout_indices(
+    predictions_trace: np.ndarray, bout_length: int, positive: bool = True, eps: float = 1e-6
+) -> np.ndarray:
+    """Find indices where a bout of bout-length occurs in a binary vector
 
     Bouts are defined as consecutive sets of 1s (if `positive`) or 0s (if not `positive`).
     Parameters
@@ -119,7 +117,7 @@ def find_bout_indices(predictions_trace: np.ndarray,
     filt = np.concatenate([[-bout_length / 2], center, [-bout_length / 2]])
     if not positive:
         predictions_trace = np.logical_not(predictions_trace.copy()).astype(int)
-    out = np.convolve(predictions_trace, filt, mode='same')
+    out = np.convolve(predictions_trace, filt, mode="same")
     # precision issues: using == 1 here has false negatives in case where out = 0.99999999998 or something
     indices = np.where(np.abs(out - 1) < eps)[0]
     if len(indices) == 0:
@@ -135,7 +133,7 @@ def find_bout_indices(predictions_trace: np.ndarray,
 
 
 def remove_short_bouts_from_trace(predictions_trace: np.ndarray, bout_length: int) -> np.ndarray:
-    """ Removes bouts of length <= `bout_length` from a binary vector.
+    """Removes bouts of length <= `bout_length` from a binary vector.
 
     Important note: we first remove "false negatives." e.g. if `bout_length` is 2, this will do something like:
         000111001111111011000010 ->  000111111111111111000010
@@ -154,7 +152,7 @@ def remove_short_bouts_from_trace(predictions_trace: np.ndarray, bout_length: in
     predictions_trace: np.ndarray. shape (N, )
         binary predictions trace with short bouts removed
     """
-    assert len(predictions_trace.shape) == 1, 'only 1D input: {}'.format(predictions_trace.shape)
+    assert len(predictions_trace.shape) == 1, "only 1D input: {}".format(predictions_trace.shape)
     # first remove 1 frame bouts, then 2 frames, then 3 frames
     for bout_len in range(1, bout_length + 1):
         # first, remove "false negatives", like filling in gaps in true behavior bouts
@@ -167,7 +165,7 @@ def remove_short_bouts_from_trace(predictions_trace: np.ndarray, bout_length: in
 
 
 def remove_short_bouts(predictions: np.ndarray, bout_length: int) -> np.ndarray:
-    """ Removes short bouts from a predictions array
+    """Removes short bouts from a predictions array
 
     Applies `remove_short_bouts_from_trace` to each column of the input.
 
@@ -183,8 +181,9 @@ def remove_short_bouts(predictions: np.ndarray, bout_length: int) -> np.ndarray:
     predictions: np.ndarray, shape (N, K)
         Array of N timepoints and K classes with short bouts removed
     """
-    assert len(predictions.shape) == 2, \
-        '2D input to remove short bouts required (timepoints x classes): {}'.format(predictions.shape)
+    assert len(predictions.shape) == 2, "2D input to remove short bouts required (timepoints x classes): {}".format(
+        predictions.shape
+    )
 
     T, K = predictions.shape
     for k in range(K):
@@ -193,7 +192,7 @@ def remove_short_bouts(predictions: np.ndarray, bout_length: int) -> np.ndarray:
 
 
 def compute_background(predictions: np.ndarray) -> np.ndarray:
-    """ Makes the background positive when no other behaviors are occurring
+    """Makes the background positive when no other behaviors are occurring
 
     Parameters
     ----------
@@ -206,29 +205,30 @@ def compute_background(predictions: np.ndarray) -> np.ndarray:
         Binary predictions. the background class is now the logical_not of whether or not there are any positive
         examples in the rest of the row
     """
-    assert len(predictions.shape) == 2, 'predictions must be a TxK matrix: not {}'.format(predictions.shape)
+    assert len(predictions.shape) == 2, "predictions must be a TxK matrix: not {}".format(predictions.shape)
 
     predictions[:, 0] = np.logical_not(np.any(predictions[:, 1:], axis=1)).astype(np.uint8)
     return predictions
 
 
 class Postprocessor:
-    """ Base class for postprocessing a set of input probabilities into predictions """
+    """Base class for postprocessing a set of input probabilities into predictions"""
+
     def __init__(self, thresholds: np.ndarray, min_threshold=0.01):
-        assert len(thresholds.shape) == 1, 'thresholds must be 1D array, not {}'.format(thresholds.shape)
+        assert len(thresholds.shape) == 1, "thresholds must be 1D array, not {}".format(thresholds.shape)
         # edge case with poor thresholds, causes all predictions to be ==1
         thresholds = remove_low_thresholds(thresholds, minimum=min_threshold)
         self.thresholds = thresholds
 
     def threshold(self, probabilities: np.ndarray) -> np.ndarray:
-        """ Applies thresholds to binarize inputs """
-        assert len(probabilities.shape) == 2, 'probabilities must be a TxK matrix: not {}'.format(probabilities.shape)
+        """Applies thresholds to binarize inputs"""
+        assert len(probabilities.shape) == 2, "probabilities must be a TxK matrix: not {}".format(probabilities.shape)
         assert probabilities.shape[1] == self.thresholds.shape[0]
         predictions = (probabilities > self.thresholds).astype(int)
         return predictions
 
     def process(self, probabilities: np.ndarray) -> np.ndarray:
-        """ Process probabilities. Will be overridden by subclasses """
+        """Process probabilities. Will be overridden by subclasses"""
         # the simplest form of postprocessing is just thresholding and making sure that background is the actual
         # logical_not of any other behavior. Therefore, its threshold is not used
         predictions = self.threshold(probabilities)
@@ -240,7 +240,8 @@ class Postprocessor:
 
 
 class MinBoutLengthPostprocessor(Postprocessor):
-    """ Postprocessor that removes bouts of length less than or equal to bout_length """
+    """Postprocessor that removes bouts of length less than or equal to bout_length"""
+
     def __init__(self, thresholds: np.ndarray, bout_length: int, **kwargs):
         super().__init__(thresholds, **kwargs)
         self.bout_length = bout_length
@@ -253,7 +254,8 @@ class MinBoutLengthPostprocessor(Postprocessor):
 
 
 class MinBoutLengthPerBehaviorPostprocessor(Postprocessor):
-    """ Postprocessor that removes bouts of length less than or equal to bout_length """
+    """Postprocessor that removes bouts of length less than or equal to bout_length"""
+
     def __init__(self, thresholds: np.ndarray, bout_lengths: list, **kwargs):
         super().__init__(thresholds, **kwargs)
         assert len(thresholds) == len(bout_lengths)
@@ -296,27 +298,25 @@ def get_bout_length_percentile(label_list: list, percentile: float) -> dict:
         bouts = get_bouts(label)
         T, K = label.shape
         for k in range(K):
-            bout_length = bouts[k]['lengths'].tolist()
+            bout_length = bouts[k]["lengths"].tolist()
             bout_lengths[k].append(bout_length)
     bout_lengths = {behavior: np.concatenate(value) for behavior, value in bout_lengths.items()}
-    # print(bout_lengths)
     percentiles = {}
     for behavior, value in bout_lengths.items():
         if len(value) > 0:
             percentiles[behavior] = np.percentile(value, percentile)
         else:
             percentiles[behavior] = 1
-    # percentiles = {behavior: np.percentile(value, percentile) for behavior, value in bout_lengths.items()}
     return percentiles
 
 
 def get_postprocessor_from_cfg(cfg: DictConfig, thresholds: np.ndarray) -> Type[Postprocessor]:
-    """ Returns a PostProcessor from an OmegaConf DictConfig returned by a  """
+    """Returns a PostProcessor from an OmegaConf DictConfig returned by a"""
     if cfg.postprocessor.type is None:
         return Postprocessor(thresholds)
-    elif cfg.postprocessor.type == 'min_bout':
+    elif cfg.postprocessor.type == "min_bout":
         return MinBoutLengthPostprocessor(thresholds, cfg.postprocessor.min_bout_length)
-    elif cfg.postprocessor.type == 'min_bout_per_behavior':
+    elif cfg.postprocessor.type == "min_bout_per_behavior":
         if not os.path.isdir(cfg.project.data_path):
             cfg = projects.convert_config_paths_to_absolute(cfg)
         assert os.path.isdir(cfg.project.data_path)
@@ -325,7 +325,7 @@ def get_postprocessor_from_cfg(cfg: DictConfig, thresholds: np.ndarray) -> Type[
         label_list = []
 
         for animal, record in records.items():
-            labelfile = record['label']
+            labelfile = record["label"]
             if labelfile is None:
                 continue
             label = file_io.read_labels(labelfile)
@@ -335,7 +335,7 @@ def get_postprocessor_from_cfg(cfg: DictConfig, thresholds: np.ndarray) -> Type[
             label_list.append(label)
 
         percentiles = get_bout_length_percentile(label_list, cfg.postprocessor.min_bout_length)
-        # percntiles is a dict: keys are behaviors, values are percentiles
+        # percentiles is a dict: keys are behaviors, values are percentiles
         # need to round and then cast to int
         percentiles = np.round(np.array(list(percentiles.values()))).astype(int)
         return MinBoutLengthPerBehaviorPostprocessor(thresholds, percentiles)
@@ -350,7 +350,7 @@ def postprocess_and_save(cfg: DictConfig) -> None:
     ----------
     cfg : DictConfig
         a project configuration. Must have the `sequence` and `postprocessing` sections
-        
+
     Goes through each "outputfile" in the project, loads the probabilities, postprocesses them, and saves to disk
     with the name `base + _predictions.csv`.
     """
@@ -361,15 +361,15 @@ def postprocess_and_save(cfg: DictConfig) -> None:
         output_name = cfg.sequence.output_name
 
     behavior_names = OmegaConf.to_container(cfg.project.class_names)
-    records = projects.get_records_from_datadir(os.path.join(cfg.project.path, 'DATA'))
+    records = projects.get_records_from_datadir(os.path.join(cfg.project.path, "DATA"))
     for _, record in records.items():
-        with h5py.File(record['output'], 'r') as f:
-            p = f[output_name]['P'][:]
-            thresholds = f[output_name]['thresholds'][:]
+        with h5py.File(record["output"], "r") as f:
+            p = f[output_name]["P"][:]
+            thresholds = f[output_name]["thresholds"][:]
             postprocessor = get_postprocessor_from_cfg(cfg, thresholds)
 
             predictions = postprocessor(p)
             df = pd.DataFrame(data=predictions, columns=behavior_names)
-            base = os.path.splitext(record['rgb'])[0]
-            filename = base + '_predictions.csv'
+            base = os.path.splitext(record["rgb"])[0]
+            filename = base + "_predictions.csv"
             df.to_csv(filename)
