@@ -1,22 +1,24 @@
 import os
 import shutil
+import time
+import platform
 
 # from projects import get_records_from_datadir, fix_config_paths
-from deepethogram import projects, utils
+from deepethogram import projects
 
 this_path = os.path.abspath(__file__)
 test_path = os.path.dirname(this_path)
 deg_path = os.path.dirname(test_path)
 
-test_data_path = os.path.join(test_path, 'DATA')
+test_data_path = os.path.join(test_path, "DATA")
 # the deepethogram test archive should only be read from, never written to
-archive_path = os.path.join(test_data_path, 'testing_deepethogram_archive')
-assert os.path.isdir(archive_path), '{} does not exist!'.format(archive_path)
-project_path = os.path.join(test_data_path, 'testing_deepethogram')
-data_path = os.path.join(project_path, 'DATA')
+archive_path = os.path.join(test_data_path, "testing_deepethogram_archive")
+assert os.path.isdir(archive_path), "{} does not exist!".format(archive_path)
+project_path = os.path.join(test_data_path, "testing_deepethogram")
+data_path = os.path.join(project_path, "DATA")
 
-config_path = os.path.join(project_path, 'project_config.yaml')
-config_path_archive = os.path.join(archive_path, 'project_config.yaml')
+config_path = os.path.join(project_path, "project_config.yaml")
+config_path_archive = os.path.join(archive_path, "project_config.yaml")
 # config_path = os.path.join(project_path, 'project_config.yaml')
 cfg_archive = projects.get_config_from_path(archive_path)
 
@@ -26,7 +28,27 @@ def change_to_deepethogram_directory():
 
 
 def clean_test_data():
-    if os.path.isdir(project_path):
+    if not os.path.isdir(project_path):
+        return
+
+    # On Windows, we need to handle file permission errors
+    if platform.system() == 'Windows':
+        max_retries = 3
+        for i in range(max_retries):
+            try:
+                shutil.rmtree(project_path)
+                break
+            except PermissionError:
+                if i < max_retries - 1:
+                    time.sleep(1)  # Wait a bit for file handles to be released
+                    continue
+                else:
+                    # If we still can't delete after retries, try to ignore errors
+                    try:
+                        shutil.rmtree(project_path, ignore_errors=True)
+                    except:
+                        pass  # If we still can't delete, just continue
+    else:
         shutil.rmtree(project_path)
 
 
@@ -39,10 +61,10 @@ def make_project_from_archive():
     # projects.fix_config_paths(cfg)
 
 
-def get_records(origin='project'):
-    if origin == 'project':
+def get_records(origin="project"):
+    if origin == "project":
         return projects.get_records_from_datadir(data_path)
-    elif origin == 'archive':
-        return projects.get_records_from_datadir(os.path.join(archive_path, 'DATA'))
+    elif origin == "archive":
+        return projects.get_records_from_datadir(os.path.join(archive_path, "DATA"))
     else:
         raise NotImplementedError
